### 2025-08-14 #8800 @editorialbot
[PRE REVIEW]: Finitewave: a lightweight and accessible framework for cardiac electrophysiology simulations
**Submitting author:** @TiNezlobinsky (<a href="http://orcid.org/0000-0002-0009-1542">Timur</a>)
**Repository:** https://github.com/finitewave/Finitewave
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.8.5
