### 2025-08-15 #8800 @Kevin-Mattheus-Moerman
@TiNezlobinsky - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
