### 2025-09-10 548a174
Add test for IsotropicStencil2D
