### 2025-09-11 16153b9
Add test for IsotropicStencil3D
