+++ tests/test_stencis.py
@@ -2,61 +2,8 @@ import numpy as np
 
 import finitewave as fw
 
-def generate_patterns():
-    mesh = []
-    w = []
 
-    mesh.append([[1, 1, 1],
-            [1, 1, 1],
-            [1, 1, 1]])
-    
-    w.append([1, 1, -4, 1, 1])
-
-    mesh.append([[1, 2, 1],
-            [1, 1, 1],
-            [1, 1, 1]])
-    
-    w.append([1, 2, -4, 0, 1])
-    
-    mesh.append([[1, 1, 1],
-            [1, 1, 2],
-            [1, 1, 1]])
-    
-    w.append([2, 1, -4, 1, 0])
-    
-    mesh.append([[1, 1, 1],
-            [1, 1, 1],
-            [1, 2, 1]])
-    
-    w.append([1, 0, -4, 2, 1])
-    
-    mesh.append([[1, 1, 1],
-            [2, 1, 1],
-            [1, 1, 1]])
-    
-    w.append([0, 1, -4, 1, 2])
-    
-    mesh.append([[1, 2, 1],
-            [1, 1, 1],
-            [1, 2, 1]])
-    
-    w.append([1, 0, -2, 0, 1])
-    
-    mesh.append([[1, 1, 1],
-            [2, 1, 2],
-            [1, 1, 1]])
-    
-    w.append([0, 1, -2, 1, 0])
-    
-    mesh.append([[1, 2, 1],
-            [2, 1, 2],
-            [1, 2, 1]])
-    
-    w.append([0, 0, 0, 0, 0])
-    return mesh, w
-
-
-def test_iso_weights_2d():
+def test_isotropic_stencil_2d():
 
     class DummyModel:
         def __init__(self):
@@ -77,14 +24,198 @@ def test_iso_weights_2d():
     model = DummyModel()
     tissue = DummyTissue()
 
-    mesh, weights = generate_patterns()
+    conductivity = np.ones((5, 5))
+    conductivity[1:-1, 1:-1] = np.array([[0.4, 0.2, 1.0],
+                                         [0.4, 1.0, 1.0],
+                                         [0.4, 1.0, 1.0]])
+    
+    w = [0.6, 0.7, -3.3, 1.0, 1.0]
+
+    tissue.conductivity = conductivity
+    
+    iso_stencil = fw.IsotropicStencil2D()
+    comp = iso_stencil.compute_weights(model, tissue)
+    comp = comp[2, 2, :]
+    w = np.array(w)  * model.D_model * model.dt / model.dr ** 2
+    w[2] += 1
+    
+    msg = "weights are not equal for conductivity"
+    assert np.allclose(w, comp, atol=1e-5), msg
+
+    
+    mesh, weights = generate_2d_patterns()
 
     for m, w in zip(mesh, weights):
-        tissue.mesh[1:-1, 1:-1] = np.array(m).T
+        tissue.mesh[1:-1, 1:-1] = np.array(m)
         tissue.conductivity = 1
         iso_stencil = fw.IsotropicStencil2D()
         comp = iso_stencil.compute_weights(model, tissue)
         comp = comp[2, 2, :]
         w = np.array(w)  * model.D_model * model.dt / model.dr ** 2
         w[2] += 1
-        assert np.allclose(w, comp, atol=10-5), f"weights are not equal for {m}"
+        msg = f"weights are not equal for {m}"
[…]
