### 2025-09-11 b79469a
Add details to the Anisotropy2D tutorial
