### 2025-09-12 d0832ba
Add test for AsymmetricStencil2D
