+++ tests/test_stencis.py
@@ -20,7 +20,7 @@ def test_isotropic_stencil_2d():
                                   [0, 1, 1, 1, 0],
                                   [0, 0, 0, 0, 0]], dtype=np.int8)
             self.conductivity = 1.0
-    
+
     model = DummyModel()
     tissue = DummyTissue()
 
@@ -28,21 +28,20 @@ def test_isotropic_stencil_2d():
     conductivity[1:-1, 1:-1] = np.array([[0.4, 0.2, 1.0],
                                          [0.4, 1.0, 1.0],
                                          [0.4, 1.0, 1.0]])
-    
+
     w = [0.6, 0.7, -3.3, 1.0, 1.0]
 
     tissue.conductivity = conductivity
-    
+
     iso_stencil = fw.IsotropicStencil2D()
     comp = iso_stencil.compute_weights(model, tissue)
     comp = comp[2, 2, :]
-    w = np.array(w)  * model.D_model * model.dt / model.dr ** 2
+    w = np.array(w) * model.D_model * model.dt / model.dr ** 2
     w[2] += 1
-    
-    msg = "weights are not equal for conductivity"
-    assert np.allclose(w, comp, atol=1e-5), msg
 
-    
+    msg = "weights are not equal for conductivity"
+    np.testing.assert_allclose(w, comp, atol=1e-5, err_msg=msg)
+
     mesh, weights = generate_2d_patterns()
 
     for m, w in zip(mesh, weights):
@@ -51,10 +50,10 @@ def test_isotropic_stencil_2d():
         iso_stencil = fw.IsotropicStencil2D()
         comp = iso_stencil.compute_weights(model, tissue)
         comp = comp[2, 2, :]
-        w = np.array(w)  * model.D_model * model.dt / model.dr ** 2
+        w = np.array(w) * model.D_model * model.dt / model.dr ** 2
         w[2] += 1
         msg = f"weights are not equal for {m}"
-        assert np.allclose(w, comp, atol=1e-5), msg
+        np.testing.assert_allclose(w, comp, atol=1e-5, err_msg=msg)
 
 
 def test_isotropic_stencil_3d():
@@ -71,7 +70,7 @@ def test_isotropic_stencil_3d():
             self.mesh = np.zeros((5, 5, 5), dtype=np.int8)
             self.mesh[1:-1, 1:-1, 1:-1] = 1
             self.conductivity = np.ones((5, 5, 5), dtype=np.float32)
-    
+
     model = DummyModel()
     tissue = DummyTissue()
 
@@ -87,17 +86,16 @@ def test_isotropic_stencil_3d():
     w = [0.5, 0.6, 0.7, -4.5, 0.8, 0.9, 1.0]
 
     tissue.conductivity[1:-1, 1:-1, 1:-1] = conductivity
-    
+
     iso_stencil = fw.IsotropicStencil3D()
     comp = iso_stencil.compute_weights(model, tissue)
     comp = comp[2, 2, 2, :]
-    w = np.array(w)  * model.D_model * model.dt / model.dr ** 2
+    w = np.array(w) * model.D_model * model.dt / model.dr ** 2
     w[3] += 1
-    
-    msg = "weights are not equal for conductivity"
-    assert np.allclose(w, comp, atol=1e-5), msg
 
-    
+    msg = "weights are not equal for conductivity"
+    np.testing.assert_allclose(w, comp, atol=1e-5, err_msg=msg)
+
     mesh, weights = generate_3d_patterns()
 
     for m, w in zip(mesh, weights):
[…]
