### 2025-09-12 41b4132
Add model, tissue, and stencil for elementalwave
