+++ finitewave/elementalwave/stencil/__init__.py
@@ -1,0 +1,1 @@
+from .triangle_isotropic_stencil import TriangleIsotropicStencil
