+++ finitewave/elementalwave/model/__init__.py
@@ -1,0 +1,1 @@
+from .aliev_panfilov_elem import AlievPanfilovElem
