+++ finitewave/elementalwave/tissue/__init__.py
@@ -1,0 +1,1 @@
+from .cardiac_tissue_elem import CardiacTissueElem
