+++ finitewave/elementalwave/__init__.py
@@ -1,0 +1,4 @@
+from .model import *
+from .tissue import *
+from .stencil import *
+from .stimulation import *
