+++ finitewave/elementalwave/solver/__init__.py
@@ -1,0 +1,1 @@
+from .scipy.scipy_cg_solver import ScipyCGSolver
