+++ finitewave/elementalwave/solver/scipy/__init__.py
@@ -1,0 +1,1 @@
+from .scipy_cg_solver import ScipyCGSolver
