### 2025-09-16 d06353e
Improve performance. Set single core for cg solver
