+++ finitewave/elementalwave/model/aliev_panfilov_elem.py
@@ -13,7 +13,9 @@ class AlievPanfilovElem(CardiacModel):
         self.cardiac_tissue = None
         self.stencil = None
         self.stim_sequence = None
-        self.rhs = np.ndarray
+        self._rhs = np.ndarray
+        self._u = np.ndarray
+        self._u_new = np.ndarray
         self.a = 0.1
         self.k = 8.0
         self.eps = 0.01
@@ -23,24 +25,34 @@ class AlievPanfilovElem(CardiacModel):
 
     def initialize(self):
         super().initialize()
-        self.rhs = np.zeros_like(self.u)
+        self._u = self.u[self.cardiac_tissue.myo_indexes].copy()
+        self._rhs = np.zeros_like(self._u)
+        self._u_new = self._u.copy()
         self.v = np.zeros_like(self.u)
+        self.u_new = self.u
+
+    def run(self, initialize=True, num_of_threads=None):
+        super().run(initialize, num_of_threads)
+        self.u[self.cardiac_tissue.myo_indexes] = self._u
+
+    def swap_arrays(self):
+        self._u_new, self._u = self._u, self._u_new
 
     def run_diffusion_kernel(self):
         """
         Executes the diffusion kernel computation using the current parameters
         and tissue weights.
         """
-        self.diffusion_kernel(self.u_new, self.u, self.rhs,
-                              self.weights, self.cardiac_tissue.myo_indexes)
+        self._u_new = self.diffusion_kernel(self._u_new, self._u, self._rhs,
+                                            self.weights)
 
     def run_ionic_kernel(self):
         """
         Executes the ionic kernel for the Aliev-Panfilov model.
         """
-        ionic_kernel_2d(self.u, self.v, self.rhs,
-                        self.cardiac_tissue.myo_indexes, self.dt,
-                        self.a, self.k, self.eps, self.mu1, self.mu2)
+        ionic_kernel_elems(self._u, self.v, self._rhs,
+                           self.cardiac_tissue.myo_indexes, self.dt,
+                           self.a, self.k, self.eps, self.mu1, self.mu2)
 
     def select_stencil(self, cardiac_tissue):
         if cardiac_tissue.elems.shape[1] == 3:
@@ -49,7 +61,7 @@ class AlievPanfilovElem(CardiacModel):
 
 
 @njit(parallel=True)
-def ionic_kernel_2d(u, v, rhs, indexes, dt, a, k, eps, mu1, mu2):
+def ionic_kernel_elems(u, v, rhs, indexes, dt, a, k, eps, mu1, mu2):
     """
     Computes the ionic kernel for the Aliev-Panfilov 2D model.
 
@@ -68,8 +80,11 @@ def ionic_kernel_2d(u, v, rhs, indexes, dt, a, k, eps, mu1, mu2):
     """
     for i in prange(len(indexes)):
         ind = indexes[i]
-        v[ind] += dt * calc_dv(v[ind], u[ind], a, k, eps, mu1, mu2)
-        rhs[ind] = dt * calc_rhs(u[ind], v[ind], a, k)
+        v[ind] += dt * calc_dv(v[ind], u[i], a, k, eps, mu1, mu2)
+        rhs[i] = dt * calc_rhs(u[i], v[ind], a, k)
+
+    # v[indexes] += dt * calc_dv(v[indexes], u[indexes], a, k, eps, mu1, mu2)
+    # rhs[indexes] = dt * calc_rhs(u[indexes], v[indexes], a, k)
 
 
 @njit
