+++ finitewave/elementalwave/solver/scipy/scipy_cg_solver.py
@@ -1,5 +1,6 @@
 from scipy.sparse import linalg
 import warnings
+from threadpoolctl import threadpool_limits
 
 
 class ScipyCGSolver:
@@ -15,16 +16,14 @@ class ScipyCGSolver:
     def axmy(self, a, b, dt):
         return a - dt * b
 
-    def diffusion_kernel(self, u_new, u, rhs, matrices, indexes):
+    @threadpool_limits.wrap(limits=1, user_api="blas")
+    def diffusion_kernel(self, u_new, u, rhs, matrices):
         a_matrix = matrices[0]
         mass_matrix = matrices[1]
-        b = mass_matrix.dot(u[indexes] + rhs[indexes])
-        u_new[indexes], success = linalg.cg(a_matrix,
-                                            b,
-                                            x0=u[indexes],
-                                            atol=self.atol,
-                                            rtol=self.rtol,
-                                            maxiter=self.maxiter,
-                                            M=self.preconditioner)
+        b = mass_matrix.dot(u + rhs)
+        u_new, success = linalg.cg(a_matrix, b, x0=u, atol=self.atol,
+                                   rtol=self.rtol, maxiter=self.maxiter,
+                                   M=self.preconditioner)
         if success > 0:
             warnings.warn("Diffusion kernel solution accuracy is not reached")
+        return u_new
