### 2025-09-16 d0c0bae
Add LuoRudy91 model
