+++ finitewave/elementalwave/model/luo_rudy_91_elems.py
@@ -1,0 +1,642 @@
+import numpy as np
+import math
+from numba import njit, prange
+
+from finitewave.core.model.cardiac_model import CardiacModel
+from finitewave.elementalwave.stencil.triangle_isotropic_stencil import (
+    TriangleIsotropicStencil
+)
+
+
+class LuoRudy91Elems(CardiacModel):
+    def __init__(self):
+        super().__init__()
+        self.cardiac_tissue = None
+        self.stencil = None
+        self.stim_sequence = None
+        self._rhs = np.ndarray
+        self._u = np.ndarray
+        self._u_new = np.ndarray
+        self.D_model = 0.1
+
+        self.m = np.ndarray
+        self.h = np.ndarray
+        self.j = np.ndarray
+        self.d = np.ndarray
+        self.f = np.ndarray
+        self.x = np.ndarray
+        self.cai = np.ndarray
+
+        self.state_vars = ["u", "m", "h", "j", "d", "f", "x", "cai"]
+        self.npfloat = 'float64'
+
+        # Ion Channel Conductances (mS/µF)
+        self.gna = 23.0     # Fast sodium (Na+) conductance
+        self.gsi = 0.045     # Slow inward calcium (Ca2+) conductance
+        self.gk = 0.705    # Time-dependent potassium (K+) conductance
+        self.gk1 = 0.6047   # Inward rectifier potassium (K1) conductance
+        self.gkp = 0.0183   # Plateau potassium (Kp) conductance
+        self.gb = 0.03921  # Background conductance (leak current)
+
+        # Extracellular and Intracellular Ion Concentrations (mM)
+        self.ko = 5.4      # Extracellular potassium concentration
+        self.ki = 145.0    # Intracellular potassium concentration
+        self.nai = 18.0     # Intracellular sodium concentration
+        self.nao = 140.0    # Extracellular sodium concentration
+        self.cao = 1.8      # Extracellular calcium concentration
+
+        # Physical Constants
+        self.R = 8.314      # Universal gas constant (J/(mol·K))
+        self.T = 310.0      # Temperature (Kelvin, 37°C)
+        self.F = 96.5       # Faraday constant (C/mmol)
+
+        # Ion Permeability Ratios
+        self.PR_NaK = 0.01833  # Na+/K+ permeability ratio
+        self.E_Na = 54.7857
+        self.E_K1 = -87.8789
+
+        # initial conditions
+        self.init_u = -84.5
+        self.init_m = 0.0017
+        self.init_h = 0.9832
+        self.init_j = 0.995484
+        self.init_d = 0.000003
+        self.init_f = 1.0
+        self.init_x = 0.0057
+        self.init_cai = 0.0002
+
+    def initialize(self):
+        super().initialize()
+        self.u[:] = self.init_u * np.ones_like(self.u)
+        self._u = self.u[self.cardiac_tissue.myo_indexes].copy()
+        self._rhs = np.zeros_like(self._u)
+        self._u_new = self._u.copy()
+        self.v = np.zeros_like(self.u)
+        self.u_new = self.u
+
+        self.m = self.init_m * np.ones_like(self.u)
+        self.h = self.init_h * np.ones_like(self.u)
+        self.j = self.init_j * np.ones_like(self.u)
+        self.d = self.init_d * np.ones_like(self.u)
+        self.f = self.init_f * np.ones_like(self.u)
+        self.x = self.init_x * np.ones_like(self.u)
+        self.cai = self.init_cai * np.ones_like(self.u)
+
+    def run(self, initialize=True, num_of_threads=None):
+        super().run(initialize, num_of_threads)
[…]
