+++ finitewave/core/model/cardiac_model.py
@@ -183,16 +183,19 @@ class CardiacModel(ABC):
         self.u_new, self.u = self.u, self.u_new
 
     def limit_num_of_threads(self, num_of_threads):
-        numba.set_num_threads(numba.config.NUMBA_NUM_THREADS)
+        max_num_of_threads = numba.config.NUMBA_NUM_THREADS
 
-        if num_of_threads is not None:
-            if num_of_threads > numba.config.NUMBA_NUM_THREADS:
-                warnings.warn(
-                    f"Selected number of threads ({num_of_threads}) exceeds the available threads ({numba.config.NUMBA_NUM_THREADS}). "
-                    f"Using the maximum available threads instead."
-                )
-            num_of_theads = min(num_of_threads, numba.config.NUMBA_NUM_THREADS)
-            numba.set_num_threads(num_of_theads)
+        if num_of_threads is None:
+            num_of_threads = max(1, max_num_of_threads - 1)
+
+        if num_of_threads > max_num_of_threads:
+            warnings.warn(
+                f"Selected number of threads ({num_of_threads}) exceeds the available threads ({max_num_of_threads}). "
+                f"Using the maximum available threads instead."
+            )
+            num_of_threads = min(num_of_threads, max_num_of_threads)
+
+        numba.set_num_threads(num_of_threads)
 
     def check_termination(self):
         """
