+++ finitewave/elementalwave/model/__init__.py
@@ -1,1 +1,2 @@
-from .aliev_panfilov_elem import AlievPanfilovElem
+from .aliev_panfilov_elems import AlievPanfilovElems
+from .luo_rudy_91_elems import LuoRudy91Elems
