### 2025-09-16 #8800 @piyueh
@TiNezlobinsky Upon a quick glance, here are two things at this moment that I would like to ask from you:

1. Fixing the invalid DOIs: See [this check](https://github.com/openjournals/joss-reviews/issues/8800#issuecomment-3187406519). Some records in `paper.bib` have an incorrect DOI format. The filed of `doi` does not need the prefix of `https://doi.org/`.
2. Briefly explaining the contributions of authors. From what I can tell in [this check](https://github.com/openjournals/joss-reviews/issues/8800#issuecomment-3187410463), only the first two authors have code contributions. Could you please describe the contributions of other authors? Here are the authorship policies we have at JOSS
   * https://joss.readthedocs.io/en/latest/submitting.html#authorship
   * "Author lists must be correct and complete. All listed authors must have made a contribution to the work, and all significant contributors should be included in the author list." as described [here](https://joss.theoj.org/about#ethics)

   Feel free to provide the contribution description here in this GitHub thread or in the paper directly.

Thank you.
