+++ paper/paper.bib
@@ -33,7 +33,7 @@
   pages    = {106223},
   year     = {2021},
   issn     = {0169-2607},
-  doi      = {https://doi.org/10.1016/j.cmpb.2021.106223},
+  doi      = {10.1016/j.cmpb.2021.106223},
   url      = {https://www.sciencedirect.com/science/article/pii/S0169260721002972},
   author   = {Gernot Plank and Axel Loewe and Aurel Neic and Christoph Augustin and Yung-Lin Huang and Matthias A.F. Gsell and Elias Karabelas and Mark Nothstein and Anton J. Prassl and Jorge Sánchez and Gunnar Seemann and Edward J. Vigmond},
   abstract = {Background and Objective: Cardiac electrophysiology is a medical specialty with a long and rich tradition of computational modeling. Nevertheless, no community standard for cardiac electrophysiology simulation software has evolved yet. Here, we present the openCARP simulation environment as one solution that could foster the needs of large parts of this community. Methods and Results: openCARP and the Python-based carputils framework allow developing and sharing simulation pipelines which automate in silico experiments including all modeling and simulation steps to increase reproducibility and productivity. The continuously expanding openCARP user community is supported by tailored infrastructure. Documentation and training material facilitate access to this complementary research tool for new users. After a brief historic review, this paper summarizes requirements for a high-usability electrophysiology simulator and describes how openCARP fulfills them. We introduce the openCARP modeling workflow in a multi-scale example of atrial fibrillation simulations on single cell, tissue, organ and body level and finally outline future development potential. Conclusion: As an open simulator, openCARP can advance the computational cardiac electrophysiology field by making state-of-the-art simulations accessible. In combination with the carputils framework, it offers a tailored software solution for the scientific community and contributes towards increasing use, transparency, standardization and reproducibility of in silico experiments.}
@@ -60,7 +60,7 @@
   pages={101252},
   year={2022},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.softx.2022.101252}
+  doi={10.1016/j.softx.2022.101252}
 }
 
 @article{Clerx2016Myokit,
@@ -108,7 +108,7 @@
   pages = {293-301},
   year = {1996},
   issn = {0960-0779},
-  doi = {https://doi.org/10.1016/0960-0779(95)00089-5},
+  doi = {10.1016/0960-0779(95)00089-5},
   url = {https://www.sciencedirect.com/science/article/pii/0960077995000895},
   author = {Rubin R. Aliev and Alexander V. Panfilov},
 }
