### 2025-09-16 #8800 @TiNezlobinsky
@piyueh, the last two authors indeed did not contribute code directly, but they were closely involved in the subject and methodological parts of the framework - including decisions on how the simulations should be designed, how to adapt them to real applications, and what should be considered physiologically correct results. Also, the weighted numerical scheme implemented in the framework was originally based on their ideas. This was a significant contribution to the development of Finitewave, and we feel it is only fair to include them as co-authors.
