### 2025-09-18 1b216ea
Add anistropy
