+++ finitewave/elementalwave/model/aliev_panfilov_elems.py
@@ -2,8 +2,8 @@ import numpy as np
 from numba import njit, prange
 
 from finitewave.core.model.cardiac_model import CardiacModel
-from finitewave.elementalwave.stencil.triangle_isotropic_stencil import (
-    TriangleIsotropicStencil
+from finitewave.elementalwave.stencil.triangle_stencil import (
+    TriangleStencil
 )
 
 
@@ -60,7 +60,7 @@ class AlievPanfilovElems(CardiacModel):
 
     def select_stencil(self, cardiac_tissue):
         if cardiac_tissue.elems.shape[1] == 3:
-            return TriangleIsotropicStencil()
+            return TriangleStencil()
         raise ValueError
 
 
