+++ finitewave/elementalwave/tissue/cardiac_tissue_elem.py
@@ -10,6 +10,7 @@ class CardiacTissueElem(CardiacTissue):
         self.mesh = np.ones(self.coords.shape[0])
         self.mesh_elems = np.ones(self.elems.shape[0])
         self.conductivity = 1.
+        self.fibers = None
 
     @property
     def coords(self):
@@ -36,11 +37,15 @@ class CardiacTissueElem(CardiacTissue):
         self.myo_indexes = np.unique(self.myo_elems)
 
     @property
+    def myo_elems_indexes(self):
+        myo_elems_mask = self.mesh_elems == 1
+        myo_vertex_mask = self.mesh == 1
+        myo_elems_mask &= np.all(myo_vertex_mask[self.elems], axis=1)
+        return np.flatnonzero(myo_elems_mask)
+
+    @property
     def myo_elems(self):
-        myo_elems = self.elems[self.mesh_elems == 1]
-        vertex_mask = self.mesh == 1
-        elem_mask = np.all(vertex_mask[myo_elems], axis=1)
-        return myo_elems[elem_mask]
+        return self.elems[self.myo_elems_indexes]
 
     @property
     def myo_coords(self):
