### 2025-09-18 a9a2a1a
Add NumbaCGSolver
