+++ finitewave/elementalwave/stencil/__init__.py
@@ -1,1 +1,2 @@
 from .triangle_stencil import TriangleStencil
+from .tetrahedral_stencil import TetrahedralStencil
