### 2025-09-18 272d060
Add tetrahedral stencil
