+++ finitewave/elementalwave/model/__init__.py
@@ -1,2 +1,0 @@
-from .aliev_panfilov_elems import AlievPanfilovElems
-from .luo_rudy_91_elems import LuoRudy91Elems
