+++ finitewave/fdmwave/stencil/stencil3D/__init__.py
@@ -1,0 +1,2 @@
+from .asymmetric_stencil_3d_fdm import AsymmetricStencil3DFDM
+from .isotropic_stencil_3d_fdm import IsotropicStencil3DFDM
