+++ finitewave/femwave/model/courtemanche_elems.py
@@ -1,0 +1,587 @@
+import numpy as np
+from numba import njit, prange
+
+from .cardiac_model_elems import CardiacModelElems
+
+
+class CourtemancheElems(CardiacModelElems):
+    """
+    A class to represent the Courtemanche cardiac model in 2D.
+
+    Attributes
+    ----------
+    D_model : float
+        Model specific diffusion coefficient.
+    state_vars : list of str
+        List of state variable names.
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.D_model = 0.154
+        self.state_vars = ["u", "nai", "ki", "cai", "caup", "carel", "m", "h", "j_",
+         "d", "f", "oa", "oi", "ua", "ui", "xr", "xs", "fca", "irel", "vrel", "urel", "wrel"]
+        
+        self.npfloat = 'float64'
+
+        # Model parameters
+        self.gna  = 7.8
+        self.gnab = 0.000674
+        self.gk1  = 0.09
+        self.gkr  = 0.0294
+        self.gks  = 0.129
+        self.gto  = 0.1652
+        self.gcal = 0.1238
+        self.gcab = 0.00113
+
+        self.gkur_coeff = 1
+
+        self.F = 96485.0
+        self.T = 310.0
+        self.R = 8314.0 
+
+        self.Vc   = 20100
+        self.Vj   = self.Vc*0.68      # (uL)
+        self.Vup  = self.Vj*0.06*0.92
+        self.Vrel = self.Vj*0.06*0.08
+
+        self.ibk = 0.0
+        self.cao = 1.8 # mM
+        self.nao = 140 # mM
+        self.ko  = 5.4 # mM
+
+        self.caupmax =  15     # mM/ms
+        self.kup     = 0.00092 # mM
+
+        self.kmnai   = 10
+        self.kmko    = 1.5
+        self.kmnancx = 87.5
+        self.kmcancx = 1.38
+        self.ksatncx = 0.1
+
+        self.kmcmdn = 0.00238
+        self.kmtrpn = 0.0005
+        self.kmcsqn = 0.8
+
+        self.trpnmax  = 0.07 # mM
+        self.cmdnmax  = 0.05 # mM
+        self.csqnmax  = 10.0 # mM
+        self.inacamax = 1600
+        self.inakmax  = 0.6
+        self.ipcamax  = 0.275
+        self.krel     = 30
+
+        self.iupmax = 0.005
+
+        self.kq10 = 3
+
+        # initial conditions
+        self.init_u     = -84.5
+        self.init_nai   = 11.2
+        self.init_ki    = 139
+        self.init_cai   = 0.000102
+        self.init_caup  = 1.6
+        self.init_carel = 1.1
+        self.init_m     = 0.00291
+        self.init_h     = 0.965
+        self.init_j     = 0.978
+        self.init_d     = 0.000137
+        self.init_f     = 0.999837
+        self.init_oa    = 0.000592
+        self.init_oi    = 0.9992
+        self.init_ua    = 0.003519
+        self.init_ui    = 0.9987
+        self.init_xs    = 0.0187
+        self.init_xr    = 0.0000329
+        self.init_fca   = 0.775
+        self.init_irel  = 0
[…]
