+++ finitewave/fdmwave/stencil/stencil2D/__init__.py
@@ -1,0 +1,3 @@
+from .isotropic_stencil_2d_fdm import IsotropicStencil2DFDM
+from .symmetric_stencil_2d_fdm import SymmetricStencil2DFDM
+from .asymmetric_stencil_2d_fdm import AsymmetricStencil2DFDM
