+++ finitewave/cpuwave3D/tissue/__init__.py
@@ -1,1 +1,0 @@
-from finitewave.cpuwave3D.tissue.cardiac_tissue_3d import CardiacTissue3D

[…]
