### 2025-09-22 98b06d8
Update animation 3D tracker
