+++ examples/trackers/FDM2D/animation_2d_tracker.py
@@ -67,8 +67,9 @@ stim_sequence.add_stim(fw.StimVoltageCoordFDM(31, 1, 0, n//2, 0, n))
 tracker_sequence = fw.TrackerSequence()
 animation_tracker = fw.AnimationTrackerFDM()
 animation_tracker.variable_name = "u"  # Specify the variable to track
-animation_tracker.dir_name = "anim_data"
-animation_tracker.step = 10
+animation_tracker.path = "."
+animation_tracker.animation_name = "anim_data"
+animation_tracker.step = 50
 animation_tracker.overwrite = True  # Remove existing files in dir_name
 tracker_sequence.add_tracker(animation_tracker)
 
