### 2025-09-23 2fec525
Add camera position to the Animation3DBuilder
