+++ finitewave/tools/animation_3d_builder.py
@@ -20,6 +20,7 @@ class Animation3DBuilder(Animation2DBuilder):
               format="mp4",
               scalar_name="Scalar",
               scalar_bar=False,
+              camera_position="iso",
               **kwargs):
         """Write the animation to a file.
 
@@ -29,13 +30,11 @@ class Animation3DBuilder(Animation2DBuilder):
                 Defaults to (800, 800).
             clim (list, optional): Color limits. Defaults to [0, 1].
             cmap (str, optional): Color map. Defaults to "viridis".
+            format (str, optional): Format of the animation. Defaults to "mp4".
+                Other options are "gif".
             scalar_name (str, optional): Name of the scalar field.
                 Defaults to "Scalar".
             scalar_bar (bool, optional): Show scalar bar. Defaults to False.
-            format (str, optional): Format of the animation. Defaults to "mp4".
-                Other options are "gif".
-            scalar_masked (bool, optional): If True, the scalar field is
-                assumed to correspond to the mask > 0. Defaults to False.
         """
 
         path = Path(self.path)
@@ -50,17 +49,20 @@ class Animation3DBuilder(Animation2DBuilder):
 
         scalar = self.load_scalar(files[0], self.scalar_mask)
 
-        print(np.nanmax(scalar))
-
         if self.scalar_mask is None:
             mesh = np.ones_like(scalar)
+            mesh[np.isnan(scalar)] = 0
         else:
             mesh = self.scalar_mask
 
-        mesh_builder = VisMeshBuilder3D()
-        mesh_builder.build_mesh(mesh, as_surface=False)
+        if mask is not None:
+            scalar[mask] = np.nan
 
-        pl = pv.Plotter(notebook=False, off_screen=False,
+        mesh_builder = VisMeshBuilder3D()
+        mesh_builder.build_mesh(mesh, as_surface=True)
+        grid = mesh_builder.add_scalar(scalar, scalar_name)
+
+        pl = pv.Plotter(notebook=False, off_screen=True,
                         window_size=window_size)
 
         if format == "mp4":
@@ -70,17 +72,9 @@ class Animation3DBuilder(Animation2DBuilder):
         else:
             raise ValueError("Format must be 'mp4' or 'gif'")
 
-        if mask is not None:
-            scalar[mask] = np.nan
-
-        import matplotlib.pyplot as plt
-
-        plt.imshow(scalar[:, :, 3])
-        plt.show()
-
-        grid = mesh_builder.add_scalar(scalar, scalar_name)
         pl.add_mesh(grid, scalars=scalar_name,
                     clim=clim, cmap=cmap, show_scalar_bar=scalar_bar)
+        pl.camera_position = camera_position
 
         pl.show(auto_close=False)
 
