+++ finitewave/femwave/solver/scipy/crank_nicolson_cg_solver.py
@@ -3,25 +3,25 @@ import warnings
 from threadpoolctl import threadpool_limits
 
 
-class ScipyCGSolver:
+class CrankNicolsonCGSolver:
     def __init__(self):
         self.maxiter = None
         self.rtol = 1e-5
         self.atol = 1e-6
         self.preconditioner = None
 
-    def axpy(self, a, b, dt):
-        return a + dt * b
-
-    def axmy(self, a, b, dt):
-        return a - dt * b
+    def assemble_matrices(self, stiffness_matrix, mass_matrix, dt):
+        a_lhs_matrix = mass_matrix + 0.5 * dt * stiffness_matrix
+        a_rhs_matrix = mass_matrix - 0.5 * dt * stiffness_matrix
+        return [a_lhs_matrix, a_rhs_matrix, mass_matrix]
 
     @threadpool_limits.wrap(limits=1, user_api="blas")
     def diffusion_kernel(self, u_new, u, rhs, matrices):
-        a_matrix = matrices[0]
-        mass_matrix = matrices[1]
-        b = mass_matrix @ (u + rhs)
-        u_new, success = linalg.cg(a_matrix, b, x0=u, atol=self.atol,
+        a_lhs_matrix = matrices[0]
+        a_rhs_matrix = matrices[1]
+        mass_matrix = matrices[2]
+        b = a_rhs_matrix @ u + mass_matrix @ rhs
+        u_new, success = linalg.cg(a_lhs_matrix, b, x0=u, atol=self.atol,
                                    rtol=self.rtol, maxiter=self.maxiter,
                                    M=self.preconditioner)
         if success > 0:
