+++ finitewave/fdmwave/model/__init__.py
@@ -1,1 +1,3 @@
 from .aliev_panfilov_fdm import AlievPanfilovFDM
+from .courtemanche_fdm import CourtemancheFDM
+
