+++ finitewave/femwave/__init__.py
@@ -2,3 +2,4 @@ from .model import *
 from .tissue import *
 from .stencil import *
 from .stimulation import *
+from .solver import *
