### 2025-09-29 ef0951c
Separate solver and assembler for fem
