+++ finitewave/fdmwave/cpuwave2D/diffusion/stencil/asymmetric_stencil_2d.py
@@ -1,0 +1,433 @@
+import numpy as np
+from numba import njit, prange
+
+from .stencil import Stencil
+
+
+class AsymmetricStencil2D(Stencil):
+    """
+    This class computes the weights for diffusion on a 2D using an asymmetric
+    stencil. The weights are calculated based on diffusion coefficients and
+    fiber orientations. The stencil includes 9 points: the central point and
+    8 surrounding points. The boundary conditions are Neumann with first-order
+    approximation.
+
+    Attributes
+    ----------
+    D_al : float
+        Longitudinal diffusion coefficient.
+    D_ac : float
+        Cross-sectional diffusion coefficient.
+
+    Notes
+    -----
+    The diffusion coefficients are general and should be adjusted according to
+    the specific model. These parameters only set the ratios between
+    longitudinal and cross-sectional diffusion.
+
+    The method assumes weights being used in the following order:
+
+    - ``w[i, j, 0] : (i-1, j-1)``,
+    - ``w[i, j, 1] : (i-1, j)``,
+    - ``w[i, j, 2] : (i-1, j+1)``,
+    - ``w[i, j, 3] : (i, j-1)``,
+    - ``w[i, j, 4] : (i, j)``,
+    - ``w[i, j, 5] : (i, j+1)``,
+    - ``w[i, j, 6] : (i+1, j-1)``,
+    - ``w[i, j, 7] : (i+1, j)``,
+    - ``w[i, j, 8] : (i+1, j+1)``.
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.D_al = 1
+        self.D_ac = 1/9
+
+    def compute_weights(self, model, cardiac_tissue):
+        """
+        Computes the weights for diffusion on a 2D mesh using an asymmetric
+        stencil.
+
+        Parameters
+        ----------
+        model : CardiacModel2D
+            A model object containing the simulation parameters.
+        cardiac_tissue : CardiacTissue2D
+            A 2D cardiac tissue object.
+
+        Returns
+        -------
+        np.ndarray
+            Array of weights for diffusion with the shape of (*mesh.shape, 9).
+        """
+        # convert fibrotic areas to non-tissue
+        mesh = cardiac_tissue.mesh.copy()
+        mesh[mesh != 1] = 0
+        conductivity = cardiac_tissue.conductivity
+        conductivity = conductivity * np.ones_like(mesh, dtype=model.npfloat)
+
+        fibers = cardiac_tissue.fibers
+
+        if fibers is None:
+            message = "Fibers must be provided for anisotropic diffusion."
+            raise ValueError(message)
+
+        weights = np.zeros((*mesh.shape, 9))
+        d_xx, d_xy = self.compute_half_step_diffusion(mesh, conductivity,
+                                                      fibers, 0)
+        d_yx, d_yy = self.compute_half_step_diffusion(mesh, conductivity,
+                                                      fibers, 1)
+        weights = compute_weights(weights, mesh, d_xx, d_xy, d_yx, d_yy)
+        weights = weights * model.D_model * model.dt / model.dr**2
+        weights[:, :, 4] += 1
+
+        return weights
+
+    def select_diffusion_kernel(self):
+        """
+        Returns the diffusion kernel function for anisotropic diffusion in 2D.
+
+        Returns
+        -------
+        function
+            The diffusion kernel function for anisotropic diffusion in 2D.
+        """
+        return diffusion_kernel_2d_aniso
+
[…]
