+++ finitewave/fdmwave/model/__init__.py
@@ -1,3 +1,0 @@
-from .aliev_panfilov_fdm import AlievPanfilovFDM
-from .courtemanche_fdm import CourtemancheFDM
-
