+++ finitewave/fdmwave/cpuwave2D/stimulation/__init__.py
@@ -1,0 +1,1 @@
+from .stim_voltage_coord_2d import StimVoltageCoord2D
