+++ finitewave/fdmwave/__init__.py
@@ -1,5 +1,1 @@
-from .model import *
-from .stencil import *
-from .tissue import *
-from .stimulation import *
-from .tracker import *
+from .cpuwave2D import *
