### 2025-10-02 6cd2544
Add simulation class
