+++ finitewave/fdmwave/cpuwave3D/tissue/__init__.py
@@ -1,0 +1,1 @@
+from .cardiac_tissue_3d import CardiacTissue3D
