### 2025-10-04 b7ca365
Add entry points for Finitewave-models
