+++ finitewave/gridywave/cpuwave/tissue/__init__.py
@@ -1,0 +1,1 @@
+from .cardiac_tissue_grid import CardiacTissueGrid
