+++ finitewave/elementalwave/simulation/__init__.py
@@ -1,0 +1,1 @@
+from .cardiac_simulation_fem import CardiacSimulationFEM
