+++ finitewave/gridywave/cpuwave/tracker/action_potential_grid_tracker.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 import numpy as np
 
-from .variable_tracker_2d import VariableTracker2D
+from .variable_grid_tracker import VariableGridTracker
 
 
-class ActionPotential2DTracker(VariableTracker2D):
+class ActionPotentialGridTracker(VariableGridTracker):
     """
     A class to track and record the action potential of a specific cell in
     a 2D cardiac tissue model.
@@ -24,7 +24,7 @@ class ActionPotential2DTracker(VariableTracker2D):
 
     def __init__(self):
         """
-        Initializes the ActionPotentialTracker2D with default parameters.
+        Initializes the ActionPotentialGridTracker with default parameters.
         """
         super().__init__()
         self.var_name = "u"
