+++ finitewave/gridywave/cpuwave3D/diffusion/stencil/asymmetric_stencil_3d.py
@@ -1,458 +1,0 @@
-import numpy as np
-from numba import njit, prange
-
-from finitewave.gridywave.stencil.stencil2D.asymmetric_stencil_2d import (
-    AsymmetricStencil2D,
-    major_component,
-    minor_component
-)
-
-
-class AsymmetricStencil3D(AsymmetricStencil2D):
-    """
-    This class computes the weights for diffusion on a 3D using an asymmetric
-    stencil. The weights are calculated based on the diffusion coefficients
-    and the fibers orientations. The stencil includes 19 points: the central
-    point and the 18 neighbors. The boundary conditions are Neumann with first-
-    order approximation.
-
-    Notes
-    -----
-    The diffusion coefficients are general and should be adjusted according to
-    the specific model. The parameters ``D_ac``, ``D_al`` only set the ratios
-    between longitudinal and cross-sectional diffusion.
-
-    The method assumes weights being used in the following order:
-        ``w[i, j, k, 0] : (i-1, j-1, k)``,
-        ``w[i, j, k, 1] : (i-1, j, k)``,
-        ``w[i, j, k, 2] : (i-1, j+1, k)``,
-        ``w[i, j, k, 3] : (i, j-1, k)``,
-        ``w[i, j, k, 4] : (i, j, k)``,
-        ``w[i, j, k, 5] : (i, j+1, k)``,
-        ``w[i, j, k, 6] : (i+1, j-1, k)``,
-        ``w[i, j, k, 7] : (i+1, j, k)``,
-        ``w[i, j, k, 8] : (i+1, j+1, k)``,
-        ``w[i, j, k, 9] : (i, j-1, k-1)``,
-        ``w[i, j, k, 10] : (i, j-1, k+1)``,
-        ``w[i, j, k, 11] : (i, j, k-1)``,
-        ``w[i, j, k, 12] : (i, j, k+1)``,
-        ``w[i, j, k, 13] : (i, j+1, k-1)``,
-        ``w[i, j, k, 14] : (i, j+1, k+1)``,
-        ``w[i, j, k, 15] : (i-1, j, k-1)``,
-        ``w[i, j, k, 16] : (i+1, j, k-1)``,
-        ``w[i, j, k, 17] : (i-1, j, k+1)``,
-        ``w[i, j, k, 18] : (i+1, j, k+1)``.
-    """
-
-    def __init__(self):
-        super().__init__()
-
-    def select_diffusion_kernel(self):
-        """
-        Selects the diffusion kernel for 3D diffusion.
-
-        Returns
-        -------
-        function
-            The diffusion kernel for 3D diffusion.
-        """
-        return diffusion_kernel_3d_aniso
-
-    def compute_weights(self, model, cardiac_tissue):
-        """
-        Computes the weights for diffusion on a 3D mesh using an asymmetric 
-        stencil.
-
-        Parameters
-        ----------
-        model : CardiacModel3D
-            A model object containing the simulation parameters.
-        cardiac_tissue : CardiacTissue3D
-            A 3D cardiac tissue object.
-
-        Returns
-        -------
-        np.ndarray
-            Array of weights for diffusion, with the shape of (*mesh.shape, 19)
-        """
-        mesh = cardiac_tissue.mesh.copy()
-        mesh[mesh != 1] = 0
-        conductivity = cardiac_tissue.conductivity
-        conductivity = conductivity * np.ones_like(mesh, dtype=model.npfloat)
-
-        fibers = cardiac_tissue.fibers
-
-        if fibers is None:
[…]
