### 2025-10-09 ef17e40
Add calc_where option for conditional equations
