### 2025-10-10 01ea5f4
Replace diffusion kernel with sparse solver
