+++ finitewave/gridywave/cpuwave/model/courtemanche.py
@@ -1,4 +1,7 @@
 from numba import njit, prange
+import jax.numpy as jnp
+import jax
+import numpy as np
 
 from .cardiac_grid_model import CardiacGridModel
 from ._registry import load_ops
@@ -124,8 +127,8 @@ class Courtemanche(CardiacGridModel):
         dt : float
             Time step for the simulation.
         """
-        ionic_kernel(u, rhs, indexes, self.myo_indexes, dt, self.nai, self.ki,
-                     self.cai, self.caup, self.carel, self.m, self.h, self.j_,
+        res = ionic_kernel(u, rhs, indexes, self.myo_indexes, dt, self.nai, self.ki,
+                           self.cai, self.caup, self.carel, self.m, self.h, self.j_,
                      self.d, self.f, self.oa, self.oi, self.ua, self.ui,
                      self.xr, self.xs, self.fca, self.irel, self.vrel,
                      self.urel, self.wrel,
@@ -139,8 +142,15 @@ class Courtemanche(CardiacGridModel):
                      self.inacamax, self.inakmax, self.ipcamax, self.krel,
                      self.iupmax, self.kq10)
 
+        (self.nai, self.ki, self.cai, self.caup, self.carel, self.m, self.h,
+         self.j_, self.d, self.f, self.oa, self.oi, self.ua, self.ui,
+         self.xr, self.xs, self.fca, self.irel, self.vrel,
+         self.urel, self.wrel) = res[1:]
 
-@njit(parallel=True)
+        return np.array(res[0])
+
+
+@jax.jit
 def ionic_kernel(u, rhs, diffusion_indexes, reaction_indexes, dt,
                  nai, ki, cai, caup, carel, m, h, j_, d, f, oa, oi, ua, ui, xs,
                  xr, fca, irel, vrel, urel, wrel,
@@ -172,76 +182,45 @@ def ionic_kernel(u, rhs, diffusion_indexes, reaction_indexes, dt,
     ... : float
         Model parameters
     """
+    ena, ek, eca = calc_equilibrum_potentials(nai, nao, ki, ko, cai, cao, R, T,
+                                              F, log=jnp.log, where=jnp.where)
 
-    for i in prange(diffusion_indexes.shape[0]):
-        d_i = diffusion_indexes[i]
-        r_i = reaction_indexes[i]
+    m = calc_gating_m(m, u, dt, exp=jnp.exp, where=jnp.where)
+    h = calc_gating_h(h, u, dt, exp=jnp.exp, where=jnp.where)
+    j_ = calc_gating_j(j_, u, dt, exp=jnp.exp, where=jnp.where)
 
-        ena, ek, eca = calc_equilibrum_potentials(nai.flat[r_i], nao,
-                                                  ki.flat[r_i], ko,
-                                                  cai.flat[r_i], cao, R, T, F)
+    ina = calc_ina(u, m, h, j_, gna, ena)
+    ik1 = calc_ik1(u, gk1, ek, exp=jnp.exp)
+    ito, oa, oi = calc_ito(u, dt, kq10, oa, oi, gto, ek, exp=jnp.exp)
+    ikur, ua, ui = calc_ikur(u, dt, kq10, ua, ui, ek, gkur_coeff, exp=jnp.exp)
+    ikr, xr = calc_ikr(u, dt, xr, gkr, ek, exp=jnp.exp)
+    iks, xs = calc_iks(u, dt, xs, gks, ek, exp=jnp.exp, sqrt=jnp.sqrt)
+    ical, d, f, fca = calc_ical(u, dt, d, f, cai, gcal, fca, exp=jnp.exp)
+    inak = calc_inak(inakmax, nai, nao, ko, kmnai, kmko, F, u, R, T,
+                     exp=jnp.exp)
[…]
