+++ finitewave/gridywave/cpuwave/diffusion/solver/__init__.py
@@ -1,0 +1,1 @@
+from .explicit_euler import ExplicitEuler
