+++ finitewave/gridywave/cpuwave/diffusion/stencil3D/__init__.py
@@ -1,2 +1,0 @@
-from .asymmetric_stencil_3d import AsymmetricStencil3D
-from .isotropic_stencil_3d import IsotropicStencil3D

[…]
