### 2025-10-11 34df102
Fix problem with diffusion kernel performance
