### 2025-10-12 ccca987
Add AsymmetricStencil for sparse solver
