+++ finitewave/gridywave/cpuwave/diffusion/stencil/asymmetric_stencil.py
@@ -1,0 +1,277 @@
+import numpy as np
+from scipy import sparse as sp
+from .stencil import Stencil
+
+
+class AsymmetricStencil(Stencil):
+    """
+    This class computes the weights for diffusion on a 2D using an isotropic
+    stencil. The stencil includes 5 points: the central point and the
+    four neighbors.
+
+    The method assumes weights being used in the following order:
+
+    - ``w[i, j, 0] : (i-1, j)``,
+    - ``w[i, j, 1] : (i, j-1)``,
+    - ``w[i, j, 2] : (i, j)``,
+    - ``w[i, j, 3] : (i, j+1)``,
+    - ``w[i, j, 4] : (i-1, j)``.
+
+    Notes
+    -----
+    The method can handle heterogeneity in the diffusion coefficients given
+    by the ``conductivity`` parameter.
+    """
+
+    def __init__(self):
+        self.boundary = IsotropicFirstOrderBoundary()
+
+    def assemble_matrices(self, simulation):
+        """
+        Computes the weights for isotropic diffusion in 2D.
+
+        Parameters
+        ----------
+        simulation : simulation
+            A simulation object containing the simulation parameters.
+
+        Returns
+        -------
+        numpy.ndarray
+            The weights for isotropic diffusion in 2D.
+        """
+        tissue = simulation.cardiac_tissue
+        model = simulation.cardiac_model
+
+        d_model = model.D_model
+        mesh = tissue.mesh.copy()
+        mesh[mesh != 1] = 0
+
+        conductivity = tissue.conductivity
+        conductivity *= np.ones_like(mesh, dtype=simulation.npfloat)
+        diffusion = d_model * conductivity / simulation.dr**2
+
+        stiff, mass = self.compute_weights_sparse(mesh, diffusion,
+                                                  tissue.myo_indexes)
+        return stiff, mass
+
+    def compute_weights_sparse(self, mesh, diffusion, indexes):
+        rows, cols, weights = self.boundary.compute(mesh, indexes)
+        weights = weights.astype(diffusion.dtype)
+        weights *= 0.5 * (diffusion.flat[cols] + diffusion.flat[rows])
+
+        c_indexes = np.zeros(mesh.size, dtype=np.int64)
+        c_indexes[indexes] = np.arange(len(indexes))
+        rows = c_indexes[rows]
+        cols = c_indexes[cols]
+
+        size = len(indexes)
+        shape = (size, size)
+        K = sp.csr_matrix((weights, (rows, cols)), shape=shape)
+        row_sums = np.array(K.sum(axis=1)).ravel()
+        D = sp.diags(-row_sums, offsets=0, format='csr')
+        K_new = K + D
+        M = sp.diags(np.ones_like(row_sums), offsets=0, format='csr')
+        return K_new.tocsr(), M.tocsr()
+
+
+class AsymmetricFirstOrderBoundary:
+    def __init__(self):
+        pass
+
+    def compute_weights(self, mesh, diffusion, indexes):
+        rows, cols, weights = [], [], []
+        ijk = np.array(np.unravel_index(indexes, mesh.shape))
+        for i in range(mesh.ndim):
+            axis = np.roll(np.arange(mesh.ndim), i)
+            major_axis = axis[0]
+            minor_axes = axis[1:]
+            for minor_axis in minor_axes:
+                d_major = diffusion[*ijk, major_axis, major_axis]
+                d_minor = diffusion[*ijk, major_axis, minor_axis]
+                for major_neighbor in [-1, 1]:
+                    res = self.compute_flow_weights(d_major, d_minor,
+                                                    major_axis, minor_axis,
+                                                    major_neighbor, mesh,
+                                                    indexes)
+                    rows += res[0]
[…]
