### 2025-10-12 438bf49
Introduce AnisotropicStencil
