+++ finitewave/gridywave/cpuwave/diffusion/stencil/anisotropic_stencil.py
@@ -3,28 +3,30 @@ from scipy import sparse as sp
 from .stencil import Stencil
 
 
-class AsymmetricStencil(Stencil):
+class AnisotropicStencil(Stencil):
     """
-    This class computes the weights for diffusion on a 2D using an isotropic
-    stencil. The stencil includes 5 points: the central point and the
-    four neighbors.
+    This class computes the weights for diffusion kernel on 2D and 3D grids
+    using an asymmetric stencil. The stencil includes 8 neighbors in 2D and
+    18 neighbors in 3D.
 
-    The method assumes weights being used in the following order:
-
-    - ``w[i, j, 0] : (i-1, j)``,
-    - ``w[i, j, 1] : (i, j-1)``,
-    - ``w[i, j, 2] : (i, j)``,
-    - ``w[i, j, 3] : (i, j+1)``,
-    - ``w[i, j, 4] : (i-1, j)``.
+    Attributes
+    ----------
+    boundary : StencilBoundary
+        An instance of the StencilBoundary class to handle Neumann boundary
+        conditions.
 
     Notes
     -----
-    The method can handle heterogeneity in the diffusion coefficients given
-    by the ``conductivity`` parameter.
+    - The method can handle heterogeneity in the diffusion coefficients given
+        by the ``conductivity`` parameter.
+    - the stencil reduces to ``IsotropicStencil`` if
+        (1) the fiber orientation is aligned with the grid,
+        (2) the `D_al = D_ac`, and
+        (3) `AsymmetricStencilBoundary` is set.
     """
 
     def __init__(self):
-        self.boundary = IsotropicFirstOrderBoundary()
+        self.boundary = AsymmetricStencilBoundary()
 
     def assemble_matrices(self, simulation):
         """
@@ -37,8 +39,10 @@ class AsymmetricStencil(Stencil):
 
         Returns
         -------
-        numpy.ndarray
-            The weights for isotropic diffusion in 2D.
+        scipy.sparse.csr_matrix
+            The stiffness matrix for the asymmetric stencil.
+        scipy.sparse.csr_matrix
+            The mass matrix for the asymmetric stencil.
         """
         tissue = simulation.cardiac_tissue
         model = simulation.cardiac_model
@@ -49,37 +53,170 @@ class AsymmetricStencil(Stencil):
 
         conductivity = tissue.conductivity
         conductivity *= np.ones_like(mesh, dtype=simulation.npfloat)
-        diffusion = d_model * conductivity / simulation.dr**2
+
+        diffusion = self.compute_diffusion(mesh, tissue.fibers,
+                                           tissue.D_al, tissue.D_ac)
+        diffusion *= (d_model * conductivity[..., np.newaxis, np.newaxis] /
+                      simulation.dr ** 2)
 
         stiff, mass = self.compute_weights_sparse(mesh, diffusion,
                                                   tissue.myo_indexes)
         return stiff, mass
 
     def compute_weights_sparse(self, mesh, diffusion, indexes):
-        rows, cols, weights = self.boundary.compute(mesh, indexes)
-        weights = weights.astype(diffusion.dtype)
-        weights *= 0.5 * (diffusion.flat[cols] + diffusion.flat[rows])
+        """
+        Computes the weights for the asymmetric stencil as a sparse matrix.
 
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        diffusion : numpy.ndarray
+            The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        indexes : numpy.ndarray
+            The indexes of the non-empty cells in the mesh.
+
+        Returns
+        -------
+        scipy.sparse.csr_matrix
+            The stiffness matrix for the asymmetric stencil.
+        scipy.sparse.csr_matrix
[…]
