+++ finitewave/gridywave/cpuwave/diffusion/__init__.py
@@ -1,3 +1,4 @@
 from .diffusion_grid_model import DiffusionGridModel
 from .stencil.isotropic_stencil import IsotropicStencil
+from .stencil.anisotropic_stencil import AnisotropicStencil
 from .solver.explicit_euler import ExplicitEuler
