### 2025-10-13 f70ab20
Add numba solvers
