+++ finitewave/elementalwave/diffusion/solver/numba/numba_linalg.py
@@ -1,0 +1,140 @@
+import numpy as np
+from numba import njit, prange
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def matvec_numba(indptr, indices, data, x, y, indexes):
+    n = indptr.shape[0] - 1
+    for i in prange(n):
+        start, end = indptr[i], indptr[i+1]
+        if start == end:
+            continue
+        y_i = 0.
+        for j in range(start, end):
+            jj = indices[j]
+            jj = indexes[jj]
+            y_i += data[j] * x.flat[jj]
+
+        ii = indexes[i]
+        y.flat[ii] = y_i
+    return y
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def matvec_and_dot_numba(indptr, indices, data, x, y, indexes):
+    n = indptr.shape[0] - 1
+    s = 0.
+    for i in prange(n):
+        start, end = indptr[i], indptr[i+1]
+        if start == end:
+            continue
+        y_i = 0.
+        for j in range(start, end):
+            jj = indices[j]
+            jj = indexes[jj]
+            y_i += data[j] * x.flat[jj]
+
+        ii = indexes[i]
+        y.flat[ii] = y_i
+        s += y_i * x.flat[ii]
+    return y, s
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def dot_numba(x, y, indexes):
+    n = len(indexes)
+    s = 0.
+    for i in prange(n):
+        ii = indexes[i]
+        s += x.flat[ii] * y.flat[ii]
+    return s
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def a_ypx_numba(a, x, y, indexes):
+    n = len(indexes)
+    for i in prange(n):
+        ii = indexes[i]
+        y.flat[ii] = a * y.flat[ii] + x.flat[ii]
+    return y
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def a_xmy_numba(a, x, y, indexes):
+    n = len(indexes)
+    for i in prange(n):
+        ii = indexes[i]
+        y.flat[ii] -= a * x.flat[ii]
+    return y
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def a_xpmy_numba(a, xp, yp, xm, ym, indexes):
+    n = len(indexes)
+    for i in prange(n):
+        ii = indexes[i]
+        yp.flat[ii] += a * xp.flat[ii]
+        ym.flat[ii] -= a * xm.flat[ii]
+    return yp, ym
+
+
+def cg_numba(indptr, indices, data, b, x, indexes, atol=1e-6, maxiter=100):
+    """ Conjugate Gradient solver for Ax = b for x, where A is given by
+    (indptr, indices, data) in CSR format.
+
+    Parameters
+    ----------
+    indptr : 1D array of int
+        CSR index pointer array.
+    indices : 1D array of int
+        CSR indices array.
+    data : 1D array of float
+        CSR data array.
+    b : 1D array of float
+        Right-hand side vector.
+    x : 1D array of float
+        Initial guess for the solution, will be modified in place.
+    indexes : 1D array of int
+        Array of indexes where the solution is defined.
+    atol : float
+        Absolute tolerance for the stopping criterion.
+    maxiter : int
+        Maximum number of iterations.
+
+    Returns
+    -------
+    x : 1D array of float
+        Approximate solution vector.
+    int
+        Number of iterations performed, or -1 if not converged within maxiter.
+    """
+
+    if np.sqrt(dot_numba(b, b, indexes)) == 0:
+        return x, 0
+
+    b0 = np.empty_like(x)
[…]
