+++ finitewave/elementalwave/diffusion/solver/scipy/crank_nicolson_cg_solver.py
@@ -8,8 +8,8 @@ from finitewave.elementalwave.diffusion.solver.solver import Solver
 class CrankNicolsonCGSolver(Solver):
     def __init__(self):
         self.maxiter = None
-        self.rtol = 1e-6
-        self.atol = 0
+        self.rtol = 0
+        self.atol = 1e-8
         self.preconditioner = None
 
     def assemble_system(self, stiffness_matrix, mass_matrix, dt):
@@ -18,7 +18,7 @@ class CrankNicolsonCGSolver(Solver):
         return [a_lhs_matrix, a_rhs_matrix, mass_matrix]
 
     @threadpool_limits.wrap(limits=1, user_api="blas")
-    def solve(self, u, rhs, matrices):
+    def solve(self, u, rhs, indexes, matrices):
         a_lhs_matrix = matrices[0]
         a_rhs_matrix = matrices[1]
         mass_matrix = matrices[2]
