### 2025-10-13 da48c0e
Join grids and elems
