+++ finitewave/gridywave/cpuwave/diffusion/stencil2D/isotropic_stencil_2d.py
@@ -1,203 +1,0 @@
-import numpy as np
-from numba import njit, prange
-
-from ..stencil import Stencil
-
-
-class IsotropicStencil2D(Stencil):
-    """
-    This class computes the weights for diffusion on a 2D using an isotropic
-    stencil. The stencil includes 5 points: the central point and the
-    four neighbors.
-
-    The method assumes weights being used in the following order:
-
-    - ``w[i, j, 0] : (i-1, j)``,
-    - ``w[i, j, 1] : (i, j-1)``,
-    - ``w[i, j, 2] : (i, j)``,
-    - ``w[i, j, 3] : (i, j+1)``,
-    - ``w[i, j, 4] : (i-1, j)``.
-
-    Notes
-    -----
-    The method can handle heterogeneity in the diffusion coefficients given
-    by the ``conductivity`` parameter.
-    """
-
-    def __init__(self):
-        super().__init__()
-
-    def diffusion_kernel(self, u_new, u, rhs, weights, indexes):
-        """
-        Returns the diffusion kernel function for isotropic diffusion in 2D.
-
-        Returns
-        -------
-        function
-            The diffusion kernel function for isotropic diffusion in 2D.
-        """
-        return diffusion_kernel(u_new, u, rhs, weights, indexes)
-
-    def compute_weights(self, simulation):
-        """
-        Computes the weights for isotropic diffusion in 2D.
-
-        Parameters
-        ----------
-        simulation : simulation
-            A simulation object containing the simulation parameters.
-
-        Returns
-        -------
-        numpy.ndarray
-            The weights for isotropic diffusion in 2D.
-        """
-        d_model = simulation.cardiac_model.D_model
-        dt = simulation.dt
-        cardiac_tissue = simulation.cardiac_tissue
-        mesh = cardiac_tissue.mesh.copy()
-        mesh[mesh != 1] = 0
-        # make sure the conductivity is a array
-        conductivity = cardiac_tissue.conductivity
-        conductivity = conductivity * np.ones_like(mesh,
-                                                   dtype=simulation.npfloat)
-        d_xx, d_yy = self.compute_half_step_diffusion(mesh, conductivity)
-
-        weights = np.zeros((*mesh.shape, 5), dtype=simulation.npfloat)
-        weights = compute_weights(weights, mesh, d_xx, d_yy,
-                                  cardiac_tissue.myo_indexes)
-        weights = weights * d_model * dt / simulation.dr**2
-        weights[:, :, 2] += 1
-
-        return weights
-
-    def compute_half_step_diffusion(self, mesh, conductivity, num_axes=2):
-        """
-        Computes the half-step diffusion values for isotropic diffusion.
-
-        Parameters
-        ----------
-        mesh : numpy.ndarray
-            A 2D array representing the mesh of the tissue.
-        conductivity : numpy.ndarray
-            A 2D array representing the conductivity of the tissue.
-        num_axes : int
-            The number of axes to compute the half-step diffusion values.
-
-        Returns
-        -------
-        numpy.ndarray
-            The half-step diffusion values for the specified axis.
-        """
-        D = np.zeros((num_axes, *mesh.shape))
-
-        for i in range(num_axes):
-            D[i] = 0.5 * (conductivity + np.roll(conductivity, -1, axis=i))
-
-        return D
-
-
-@njit(parallel=True)
-def diffusion_kernel(u_new, u, rhs, w, indexes):
-    """
-    Performs isotropic diffusion on a 2D grid.
-
[…]
