+++ finitewave/core/model/__init__.py
@@ -1,1 +1,1 @@
-from .cardiac_model import CardiacModel
+from .cardiac_model_base import CardiacModelBase
