+++ finitewave/core/model/cardiac_model_base.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 import copy
 
 
-class CardiacModel(ABC):
+class CardiacModelBase(ABC):
     """
     Base class for electrophysiological models.
 
