### 2025-10-14 a39b8cd
Fix mask condition in anisotropic_stencil.py
