+++ finitewave/cpuwave/diffusion/stencil/anisotropic_stencil.py
@@ -402,7 +402,7 @@ class AsymmetricStencilBoundary:
         m_upper = m3 + m4 + m + m0
         m_lower = m1 + m2 + m + m0
 
-        mask = ((m == 0) | (m3 == 0) | (m_upper < 3) | (m_lower < 3))
+        mask = ((m == 0) | (m0 == 0) | (m_upper < 3) | (m_lower < 3))
 
         w = np.where(mask, 0, m / m_upper - m / m_lower)
         w0 = np.where(mask, 0, m0 / m_upper - m0 / m_lower)
