### 2025-10-17 10a2e9c
Add memory_save option for StimVoltageMatrix
