### 2025-10-25 39ea9e0
Fix ForwardEulerSolver
