+++ finitewave/cpuwave/diffusion/assembler/triangle_assembler.py
@@ -1,30 +1,129 @@
 import numpy as np
 
-from .assembler import Assembler
+from .element_assembler import ElementAssembler
 
 
-class TriangleAssembler(Assembler):
+class TriangleAssembler(ElementAssembler):
     def __init__(self):
         super().__init__()
         self.mass_coef = 12.
+        self.elem_mass = (1 / self.mass_coef) * np.array([[2, 1, 1],
+                                                          [1, 2, 1],
+                                                          [1, 1, 2]])
+        # self.elem_mass = (1/3) * np.eye(3)
+        self.dN_dxi = np.array([-1.0, 1.0, 0.0])
+        self.dN_deta = np.array([-1.0, 0.0, 1.0])
+        self.quad_weights = np.array([1/2])
+        self.n_points = 3
 
     def volumes_and_grads(self, coords, elems):
-        grads = np.zeros((elems.shape[0], 3, 3))
+        """
+        Computes area and global gradients for linear triangles in 3D
+        using the Jacobian from the reference triangle.
 
-        # vertice 0
-        p0 = coords[elems[:, 0]]
-        p1 = coords[elems[:, 1]]
-        p2 = coords[elems[:, 2]]
+        Parameters:
+        ----------
+        coords: (N_nodes, 3)
+            Coordinates of the mesh nodes.
+        elems: (N_elems, 3)
+            Element connectivity (node indices for each triangle).
 
-        normals = np.cross(p1 - p0, p2 - p0)
-        areas = 0.5 * np.linalg.norm(normals, axis=1)
-
-        phi_0 = np.cross(p1 - p2, normals) / (2.0 * areas[:, np.newaxis]) ** 2
-        phi_1 = np.cross(p2 - p0, normals) / (2.0 * areas[:, np.newaxis]) ** 2
-        phi_2 = np.cross(p0 - p1, normals) / (2.0 * areas[:, np.newaxis]) ** 2
-
-        grads[:, 0, :] = phi_0
-        grads[:, 1, :] = phi_1
-        grads[:, 2, :] = phi_2
+        Returns:
+        -------
+        areas: (N_elems,)
+            Area of each triangle element.
+        grads: (N_elems, 3, 3)
+            Gradient of shape functions in global coordinates for each element.
+        """
+        jacobian = self.build_jacobian(coords, elems)
+        areas = self.compute_areas(jacobian)
+        grads = self.compute_gradients(jacobian, self.n_points)
 
         return areas, grads
+
+    def build_jacobian(self, coords, elems):
+        """
+        Build Jacobian matrices for quadrilateral elements.
+
+        Parameters:
+        ----------
+        coords: (N_nodes, 3)
+            Coordinates of the mesh nodes.
+        elems: (N_elems, 4)
+            Element connectivity (node indices for each quadrilateral).
+
+        Returns:
+        -------
+        jacobian: (N_elems, 2, 3)
+            Jacobian matrices for each quadrilateral element.
+        """
+        n_elems = elems.shape[0]
+        jacobian = np.zeros((n_elems, 2, 3))
+
+        for i in range(self.n_points):
+            jacobian[:, 0, :] += self.dN_dxi[i] * coords[elems[:, i]]
+            jacobian[:, 1, :] += self.dN_deta[i] * coords[elems[:, i]]
+
+        return jacobian
+
+    def compute_gradients(self, jacobian, n_points):
+        """Compute global gradients for triangle elements.
+
+        Parameters:
+        ----------
+        jacobian: (N, 2, 3)
+            Jacobian matrices for N triangle elements.
+
+        Returns:
+        -------
+            grads: (N, n_points, 3)
[…]
