+++ finitewave/cpuwave/diffusion/stencil/anisotropic_stencil.py
@@ -90,7 +90,7 @@ class AnisotropicStencil(Stencil):
 
         size = len(indexes)
         shape = (size, size)
-        K_stiff = sp.csr_matrix((weights, (rows, cols)), shape=shape)
+        K_stiff = - sp.csr_matrix((weights, (rows, cols)), shape=shape)
         M_mass = sp.diags(np.ones_like(indexes, dtype=weights.dtype),
                           offsets=0, format='csr')
         return K_stiff.tocsr(), M_mass.tocsr()
