+++ finitewave/cpuwave/diffusion/__init__.py
@@ -1,5 +1,3 @@
 from .diffusion_model import DiffusionModel
-from .stencil.isotropic_stencil import IsotropicStencil
-from .stencil.anisotropic_stencil import AnisotropicStencil
-from .assembler.triangle_assembler import TriangleAssembler
-from .assembler.tetrahedral_assembler import TetrahedralAssembler
+from .stencil import *
+from .assembler import *
