+++ finitewave/cpuwave/tissue/__init__.py
@@ -1,2 +1,7 @@
 from .cardiac_tissue_grid import CardiacTissueGrid
 from .cardiac_tissue_elements import CardiacTissueElements
+
+__all__ = [
+    "CardiacTissueGrid",
+    "CardiacTissueElements",
+]
