### 2025-10-29 5d9a261
Add GridAssembler
