+++ examples/basics/grid2D/isotropic_medium_2d.py
@@ -36,16 +36,22 @@ showing the resulting excitation wave pattern.
 import matplotlib.pyplot as plt
 
 import finitewave as fw
+import numpy as np
 
 # create a tissue of size 400x400 with cardiomycytes:
 n = 400
 tissue = fw.CardiacTissueGrid([n, n])
+# tissue.mesh += (np.random.random(tissue.mesh.shape) < 0.3).astype(tissue.mesh.dtype)
 
 # set up stimulation parameters:
 stim_sequence = fw.StimSequence()
 stim_sequence.add_stim(fw.StimVoltageCoord(time=0, volt_value=1,
                                            x_min=n//2 - 3, x_max=n//2 + 3,
                                            y_min=n//2 - 3, y_max=n//2 + 3))
+
+diffusion_model = fw.DiffusionModel()
+diffusion_model.assembler = fw.GridAssembler()
+diffusion_model.assembler.stencil = fw.IsotropicStencil()
 
 # create model object and set up parameters:
 simulation = fw.CardiacSimulation()
@@ -55,6 +61,7 @@ simulation.t_max = 30
 simulation.cardiac_model = fw.AlievPanfilov()
 simulation.cardiac_tissue = tissue
 simulation.stim_sequence = stim_sequence
+# simulation.diffusion_model = diffusion_model
 
 # run the model:
 simulation.run()
