+++ finitewave/cpuwave/diffusion/stencil/__init__.py
@@ -1,2 +1,0 @@
-from .isotropic_stencil import IsotropicStencil
-from .anisotropic_stencil import AnisotropicStencil
