+++ finitewave/cpuwave/diffusion/stencils/__init__.py
@@ -1,0 +1,3 @@
+from .assymetric_stencil import AsymmetricStencil
+from .isotropic_stencil import IsotropicStencil
+from .isotropic_second_order_stencil import IsotropicSecondOrderStencil
