### 2025-10-30 39f6dc6
Add SymmetricStencil
