+++ finitewave/cpuwave/diffusion/assembler/surface_assembler.py
@@ -4,6 +4,19 @@ from .element_assembler import ElementAssembler
 
 
 class SurfaceAssembler(ElementAssembler):
+    """
+    Class for assembling surface element diffusion models.
+
+    Attributes
+    ----------
+    reference_element : ReferenceElement
+        The reference element used for numerical integration.
+    simulation : Simulation
+        The simulation instance associated with this assembler.
+    weights : tuple
+        The computed diffusion weights (stiffness and mass matrices).
+    """
+
     def __init__(self):
         super().__init__()
         self.reference_element = None
