+++ finitewave/cpuwave/diffusion/diffusion_model.py
@@ -4,8 +4,7 @@ from .elements.triangle_element import LinearTriangleElement
 from .elements.quadrilateral_element import LinearQuadrilateralElement
 from .elements.tetrahedral_element import LinearTetrahedralElement
 
-from .stencils.isotropic_stencil import IsotropicStencil
-from .stencils.assymetric_stencil import AsymmetricStencil
+from .stencils.asymmetric_stencil import AsymmetricStencil
 
 from .assembler.surface_assembler import SurfaceAssembler
 from .assembler.volume_assembler import VolumeAssembler
@@ -13,24 +12,56 @@ from .assembler.grid_assembler import GridAssembler
 
 
 class DiffusionModel(DiffusionModelBase):
+    """
+    Diffusion model which selects appropriate assembler based on tissue type.
+
+    This model supports both grid-based and element-based simulations,
+    adapting the diffusion stencil and assembler accordingly.
+
+    Attributes
+    ----------
+    assembler : Assembler
+        The assembler used to construct the diffusion matrices.
+    simulation : Simulation
+        The simulation instance associated with this diffusion model.
+    weights : dict
+        The computed diffusion weights for the simulation.
+    """
     def __init__(self):
         super().__init__()
         self.assembler = None
         self.simulation = None
 
     def initialize(self, simulation):
+        """Selects and initializes the appropriate assembler and
+        computes the diffusion weights.
+
+        Parameters
+        ----------
+        simulation : Simulation
+            The simulation instance associated with this diffusion model.
+        """
         super().initialize(simulation)
         self.simulation = simulation
 
         if self.assembler is None:
             self.assembler = self.default_assembler()
+            self.assembler.initialize(simulation)
 
         self.compute_weights()
 
     def compute_weights(self):
-        self.weights = self.assembler.assemble_matrices(self.simulation)
+        """Computes weights using the selected assembler."""
+        self.weights = self.assembler.compute_weights()
 
     def default_assembler(self):
+        """Selects the default assembler based on tissue type.
+
+        Returns
+        -------
+        Assembler
+            The selected assembler instance.
+        """
         if self.simulation.cardiac_tissue.meta['type'] == 'Grid':
             return self._default_stencil()
 
