+++ finitewave/cpuwave/diffusion/stencils/symmetric_stencil.py
@@ -1,0 +1,406 @@
+import numpy as np
+
+
+class SymmetricStencil:
+    """
+    This class computes indexes and weights for the symmetric stencil.
+
+    Notes
+    -----
+    The symmetric stencil reduces to the isotropic stencil with first-order
+    boundary conditions if the ``fibers = None`` or ``fibers`` are aligned with
+    the grid and ``D_al = D_ac``.
+
+    Rules for handling boundaries are:
+    - If a major (directly adjacent) neighbor is invalid
+        (out of bounds or in an empty cell), flow from this neighbor is zero.
+    - If more than one minor neighbor from upper or lower side is invalid,
+        flow from these neighbors is determined only by the major component.
+    - If a minor neighbor from upper or lower side is invalid, minor component
+        is calculated using the remaining valid minor neighbors.
+
+    Diffusion components are calculated in the middle of two nodes, therefore
+    the diffusion coefficient is averaged between the corresponding nodes.
+
+    References
+    ----------
+    Bram van Es, Barry Koren, Hugo J. de Blank,
+    Finite-difference schemes for anisotropic diffusion,
+    Journal of Computational Physics,
+    Volume 272, 2014, Pages 526-549, ISSN 0021-9991,
+    https://doi.org/10.1016/j.jcp.2014.04.046.
+    """
+    def __init__(self):
+        pass
+
+    def compute_weights(self, mesh, diffusion, indexes):
+        """
+        Computes the weights for the asymmetric stencil.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        diffusion : numpy.ndarray
+            The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        indexes : numpy.ndarray
+            The indexes of the non-empty cells in the mesh.
+        """
+        rows, cols, weights = [], [], []
+
+        print(diffusion[0, 0, :, :])
+        for i in range(mesh.ndim):
+            axis = np.roll(np.arange(mesh.ndim), i)
+            major_axis = axis[0]
+            minor_axes = axis[1:]
+            for minor_axis in minor_axes:
+                D = diffusion[..., [major_axis, minor_axis]][..., [major_axis, minor_axis], :].copy()
+
+                for major_neighbor in [-1, 1]:
+                    res = self.compute_flow_weights(D, major_axis, minor_axis,
+                                                    major_neighbor, mesh, indexes)
+                    rows += res[0]
+                    cols += res[1]
+                    weights += res[2]
+
+        rows = np.concatenate(rows)
+        cols = np.concatenate(cols)
+        weights = np.concatenate(weights)
+        return rows, cols, weights
+
+    def compute_flow_weights(self, D, major_axis, minor_axis, major_neighbor,
+                             mesh, indexes):
+        """
+        Calculates the flow from major to center.
+
+        .. code-block:: text
[…]
