+++ finitewave/cpuwave/diffusion/assembler/volume_assembler.py
@@ -4,6 +4,18 @@ from .element_assembler import ElementAssembler
 
 
 class VolumeAssembler(ElementAssembler):
+    """
+    Class for assembling volume element diffusion models.
+
+    Attributes
+    ----------
+    reference_element : ReferenceElement
+        The reference element used for numerical integration.
+    simulation : Simulation
+        The simulation instance associated with this assembler.
+    weights : tuple
+        The computed diffusion weights (stiffness and mass matrices).
+    """
     def __init__(self):
         super().__init__()
 

[…]
