### 2025-10-31 cfa3fc2
Stable SymmetricStencil for fibrosis and update Tutorial
