+++ finitewave/cpuwave/diffusion/stencils/symmetric_stencil.py
@@ -48,7 +48,6 @@ class SymmetricStencil:
         """
         rows, cols, weights = [], [], []
 
-        print(diffusion[0, 0, :, :])
         for i in range(mesh.ndim):
             axis = np.roll(np.arange(mesh.ndim), i)
             major_axis = axis[0]
@@ -87,8 +86,6 @@ class SymmetricStencil:
         ijk_list, mask_list = self.valid_neighbors(mesh, indexes, major_axis,
                                                    minor_axis, major_neighbor)
         ijk, ijk_0, ijk_1, ijk_2, ijk_3, ijk_4 = ijk_list
-
-        print(major_axis, minor_axis, major_neighbor)
 
         m, m0, m1, m2, m3, m4 = mask_list
         d_upper = np.zeros((ijk.shape[1], 2, 2), dtype=D.dtype)
@@ -130,8 +127,8 @@ class SymmetricStencil:
                 |                   |
              minor_1 ------------ minor_2
         """
-        minor_lower = major_shift
-        minor_upper = -major_shift
+        minor_lower = - major_shift
+        minor_upper = major_shift
 
         ijk_center = np.array(np.unravel_index(indexes, mesh.shape))
         ijk_major = self.build_neighbor(ijk_center, major_shift, major_axis)
@@ -211,21 +208,25 @@ class SymmetricStencil:
         w4 = np.zeros_like(m4, dtype=d_upper.dtype)
 
         # m, m0, m3, m4
-        w_, w0_, w3_, w4_ = self.flow_upper_component(d_upper, m, m0, m3, m4)
+        mx = (-0.5, 0.5, -0.5, 0.5)
+        my = (-0.5, -0.5, 0.5, 0.5)
+        w_, w0_, w3_, w4_ = self.flow_component(d_upper, m, m0, m3, m4, mx, my)
         w += 0.5 * w_
         w0 += 0.5 * w0_
         w3 += 0.5 * w3_
         w4 += 0.5 * w4_
 
         # m, m0, m1, m2
-        w_, w0_, w1_, w2_ = self.flow_lower_component(d_lower, m, m0, m1, m2)
+        mx = (-0.5, 0.5, -0.5, 0.5)
+        my = (0.5, 0.5, -0.5, -0.5)
+        w_, w0_, w1_, w2_ = self.flow_component(d_lower, m, m0, m1, m2, mx, my)
         w += 0.5 * w_
         w0 += 0.5 * w0_
         w1 += 0.5 * w1_
         w2 += 0.5 * w2_
         return w, w0, w1, w2, w3, w4
 
-    def flow_upper_component(self, d, m0, m1, m2, m3):
+    def flow_component(self, d, m0, m1, m2, m3, mx, my):
         """
         q = d @ [dx, dy]
         q @ n = 0, with n the normal vector at the boundary
@@ -234,6 +235,9 @@ class SymmetricStencil:
         dx = 0.5 * (u1 + u3) - 0.5 * (u0 + u2)
         dy = 0.5 * (u2 + u3) - 0.5 * (u0 + u1)
 
+        dx = mx0 * u0 + mx1 * u1 + mx2 * u2 + mx3 * u3
+        dy = my0 * u0 + my1 * u1 + my2 * u2 + my3 * u3
+
         .. code-block:: text
[…]
