### 2025-11-01 42b3a63
Fix stencils for 3D
