+++ finitewave/cpuwave/diffusion/stencils/__init__.py
@@ -1,4 +1,4 @@
 from .asymmetric_stencil import AsymmetricStencil
-from .symmetric_stencil import SymmetricStencil
+from .symmetric_stencil_2d import SymmetricStencil2D
 from .isotropic_stencil import IsotropicStencil
 from .isotropic_second_order_stencil import IsotropicSecondOrderStencil
