+++ finitewave/cpuwave/stimulation/__init__.py
@@ -1,3 +1,7 @@
 from .stim_voltage_coord import StimVoltageCoord
 from .stim_voltage_electrodes import StimVoltageElectrodes
 from .stim_voltage_matrix import StimVoltageMatrix
+from .stim_voltage_collection import StimVoltageCollection
+from .stim_current_coord import StimCurrentCoord
+from .stim_current_electrodes import StimCurrentElectrodes
+from .stim_current_matrix import StimCurrentMatrix
