### 2025-11-07 7fea29a
Add sphere gif
