### 2025-11-07 2e0fbe9
Add animation builder for elems
