### 2025-11-14 TiNezlobinsky/Finitewave#20 @sitic
Missing PyPi package
It would be great if `finitewave` could be installed with

```
pip install finitewave
```

The project name [finitewave](https://pypi.org/search/?q=finitewave) looks to be available on PyPi. After a brief look at `pyproject.toml` it seems the package is already well-structured for this; the only missing step is the registration and upload to PyPI. See this [guide](https://packaging.python.org/en/latest/tutorials/packaging-projects/#uploading-the-distribution-archives) for more details.

The [JOSS review guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions) strongly recommend automated installation, ping to review issue: https://github.com/openjournals/joss-reviews/issues/9310
