### 2026-01-03 c2a1c54
Add docs for solvers. Remove unused scipy solvers
