### 2026-02-16 #9310 @TiNezlobinsky
Dear editor @piyueh  and reviewers @ashahide, @sitic.

We are currently preparing a new update of the framework, which we expect to release within a week. This update will not fundamentally change how the framework works, but it may require small adjustments to some scripts described in the paper.

If the review process has not yet started, we would be happy to release the updated version and adjust the manuscript within the coming week to ensure everything is fully aligned.

However, if the review is already underway and updating the submission would complicate the process, we are equally happy to proceed with the current version.

Please let us know what would be most convenient.

Thank you again for your time and consideration.

Timur
