### 2026-02-16 #9310 @piyueh
@TiNezlobinsky Yes, it is possible to change the version covered by the JOSS paper, provided the reviewers agree. However, this is not required. You may continue releasing new versions without modifying the JOSS submission. The decision depends on which version you would like the JOSS paper to describe.

Once published, the JOSS paper will apply exclusively to version v0.8.5 of the software. It does not extend to subsequent releases. Reviewers are evaluating only the code corresponding to v0.8.5. Therefore, later changes to the codebase or new releases will not affect the review, as long as the v0.8.5 snapshot remains available (for example, as a tagged release or branch).

In this situation, I would generally recommend keeping the version covered by the JOSS paper unchanged while continuing development and making new releases, unless there are new features or critical bug fixes that you would like reflected in the published paper.

Please let me know whether you would prefer to update the version covered by the JOSS paper or keep it at v0.8.5.

If you decide not to change the version associated with the JOSS paper while continuing to make new releases, please keep some flexibility in your version numbering. Reviewers may request changes to v0.8.5, and at the end of the review process you may need to issue a new release that includes v0.8.5 plus the requested revisions (if there's any). If version number consistency is important to you, you may wish to reserve a subsequent version number (e.g., v0.8.6 or v0.8.7) for the final JOSS-related release.

@ashahide @sitic Apologies for not checking in earlier on the review progress. If possible, could we expedite this submission? Thank you.
