### 2026-02-20 1e45abc
Extract assemblers. Restructure the modules.
