+++ finitewave/cpuwave/numerics/fdm/asymmetric_stencil.py
@@ -14,9 +14,9 @@ class AsymmetricStencil(Stencil):
 
     Rules for handling boundaries are:
     - If a major (directly adjacent) neighbor is invalid
-        (out of bounds or in an empty cell), flow from this neighbor is zero.
+        (out of bounds or in an empty cell), flux from this neighbor is zero.
     - If more than one minor neighbor from upper or lower side is invalid,
-        flow from these neighbors is determined only by the major component.
+        flux from these neighbors is determined only by the major component.
     - If a minor neighbor from upper or lower side is invalid, minor component
         is calculated using the remaining valid minor neighbors.
 
@@ -33,47 +33,10 @@ class AsymmetricStencil(Stencil):
     """
     def __init__(self):
         super().__init__()
-
-    def compute_weights(self, mesh, diffusion, indexes):
+   
+    def compute_flux_weights(self, mesh, diffusion, dr, ijk, major_axis):
         """
-        Computes the weights for the asymmetric stencil.
-
-        Parameters
-        ----------
-        mesh : numpy.ndarray
-            The mesh of the simulation.
-        diffusion : numpy.ndarray
-            The diffusion tensor as a (*mesh.shape, ndim, ndim).
-        indexes : numpy.ndarray
-            The indexes of the non-empty cells in the mesh.
-
-        Returns
-        -------
-        tuple
-            A tuple containing the rows, cols, and weights for the stencil.
-        """
-        rows, cols, weights = [], [], []
-        ijk = np.array(np.unravel_index(indexes, mesh.shape))
-        for i in range(mesh.ndim):
-            axis = np.roll(np.arange(mesh.ndim), i)
-            major_axis = axis[0]
-            minor_axes = axis[1:]
-
-            res = self.compute_flow_weights(mesh, diffusion, ijk, major_axis,
-                                            minor_axes)
-            rows += res[0]
-            cols += res[1]
-            weights += res[2]
-
-        rows = np.concatenate(rows)
-        cols = np.concatenate(cols)
-        weights = np.concatenate(weights)
-        return rows, cols, weights
-
-    def compute_flow_weights(self, mesh, diffusion, ijk, major_axis,
-                             minor_axes):
-        """
-        Computes the flow weights for the asymmetric stencil.
+        Computes the flux weights along major axis.
 
         .. code-block:: text
              minor_3 ----------- minor_4
@@ -92,98 +55,52 @@ class AsymmetricStencil(Stencil):
             The mesh of the simulation.
         diffusion : numpy.ndarray
             The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        dr : float
+            The grid spacing.
         ijk : numpy.ndarray
             The indexes of the non-empty cells in the mesh.
         major_axis : int
             The axis of the major direction.
-        minor_axes : list
-            The axes of the minor directions.
 
-        Returns
+        Returns 
         -------
         tuple
-            A tuple containing the rows, cols, and weights for the flow.
+            A tuple containing the rows, cols, and weights for the flux.
         """
-        res = self.major_flow_weights(mesh, diffusion, ijk, major_axis)
-        ijk_center, ijk_major, m_center, m_major, w_center, w_major = res
-        ijk_list = [ijk_center, ijk_major]
-        w_list = [w_center, w_major]
+        ijk_major = self.build_neighbor(ijk, shift=1, axis=major_axis)
+        m_center = self.is_valid_index(ijk, mesh)
+        m_major = self.is_valid_index(ijk_major, mesh)
+        
+        ijk_list, w_list = self.major_flux_weights(diffusion, dr,
[…]
