+++ finitewave/cpuwave/model/__init__.py
@@ -1,2 +1,3 @@
 from .aliev_panfilov import AlievPanfilov
 from .courtemanche import Courtemanche
+from .fenton_karma import FentonKarma
