+++ finitewave/cpuwave/numerics/linalg/__init__.py
@@ -1,0 +1,1 @@
+from .numba_linalg import *
