+++ finitewave/cpuwave/solver/numba/__init__.py
@@ -1,3 +1,0 @@
-from .backward_euler_cg_solver import BackwardEulerCGSolver
-from .crank_nicolson_cg_solver import CrankNicolsonCGSolver
-from .forward_euler_solver import ForwardEulerSolver
