+++ finitewave/cpuwave/solver/__init__.py
@@ -1,1 +1,0 @@
-from .numba import *

[…]
