### 2026-02-25 6f1a940
Add cell-based symmetric stencil
