+++ finitewave/cpuwave/numerics/fdm/__init__.py
@@ -1,2 +1,3 @@
 from .asymmetric_stencil import AsymmetricStencil
 from .isotropic_stencil import IsotropicStencil
+from .cell_stencil import CellStencil
