### 2026-02-27 0ea897e
Update atrial example
