### 2026-03-02 6b032ae
Add poisson solver with numba support
