### 2026-03-04 941fc09
Add preconditioner support
