+++ finitewave/cpuwave/numerics/linalg/numba_linalg.py
@@ -3,57 +3,10 @@ from numba import njit, prange
 
 
 @njit(parallel=True, fastmath=True, cache=True)
-def forward_euler(indptr, indices, data, u, rhs, mass_inv, u_new, indexes, dt):
-    """
-    Performs the Forward Euler step:
-    y = u + M^{-1} * (A * u + rhs)
-
-    Parameters
-    ----------
-    indptr : np.ndarray
-        CSR index pointer array.
-    indices : np.ndarray
-        CSR indices array.
-    data : np.ndarray
-        CSR data array.
-    u : np.ndarray
-        Current solution vector.
-    rhs : np.ndarray
-        Right-hand side vector.
-    mass_inv : np.ndarray
-        Inverse of the lumped mass matrix diagonal.
-    u_new : np.ndarray
-        Output solution vector.
-    indexes : np.ndarray
-        Array of indexes where the solution is defined.
-
-    Returns
-    -------
-    y : np.ndarray
-        Updated solution vector after the Forward Euler step.
-    """
-    n_rows = indptr.size - 1
-
-    for i in prange(n_rows):
-        start, end = indptr[i], indptr[i+1]
-        if start == end:
-            continue
-        i_tr = 0.0
-        for j in range(start, end):
-            jj = indices[j]
-            jj = indexes[jj]
-            i_tr += data[j] * u.flat[jj]
-
-        ii = indexes[i]
-        u_new.flat[ii] = (u.flat[ii] -
-                          dt * mass_inv.flat[i] * i_tr +
-                          dt * rhs.flat[ii])
-
-    return u_new
-
-
-@njit(parallel=True, fastmath=True, cache=True)
 def matvec_numba(indptr, indices, data, x, y, indexes):
+    """
+    Computes y = A @ x for a sparse matrix A in CSR format.
+    """
     n = indptr.shape[0] - 1
     for i in prange(n):
         start, end = indptr[i], indptr[i+1]
@@ -72,6 +25,8 @@ def matvec_numba(indptr, indices, data, x, y, indexes):
 
 @njit(parallel=True, fastmath=True, cache=True)
 def matvec_and_dot_numba(indptr, indices, data, x, y, indexes):
+    """Computes y = A @ x and returns the dot product x^T @ (A @ x).
+    """
     n = indptr.shape[0] - 1
     s = 0.
     for i in prange(n):
@@ -88,6 +43,27 @@ def matvec_and_dot_numba(indptr, indices, data, x, y, indexes):
         y.flat[ii] = y_i
         s += y_i * x.flat[ii]
     return y, s
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def b_m_matvec_numba(indptr, indices, data, b, x, y, indexes):
+    """
+    Computes y = b - A @ x
+    """
+    n = indptr.shape[0] - 1
+    for i in prange(n):
+        start, end = indptr[i], indptr[i+1]
+        if start == end:
+            continue
+        y_i = 0.
+        for j in range(start, end):
+            jj = indices[j]
+            jj = indexes[jj]
+            y_i += data[j] * x.flat[jj]
+
+        ii = indexes[i]
+        y.flat[ii] = b.flat[ii] - y_i
+    return y
 
 
 @njit(parallel=True, fastmath=True, cache=True)
@@ -133,69 +109,4 @@ def y_pm_ax_numba(a, xp, yp, xm, ym, indexes):
         yp.flat[ii] += a * xp.flat[ii]
         ym.flat[ii] -= a * xm.flat[ii]
     return yp, ym
-
-
-def cg_numba(indptr, indices, data, b, x, indexes, rtol=None, atol=1e-6, maxiter=100):
-    """ Conjugate Gradient solver for Ax = b for x, where A is given by
[…]
