+++ finitewave/cpuwave/numerics/linalg/__init__.py
@@ -1,2 +1,2 @@
-from .numba_linalg import *
 from .poisson_cg_solver import poisson_cg_solver
+from .preconditioner import JacobiPreconditioner
