### 2026-03-05 bae6fb0
Add prepacing method and fix courtemanche model
