+++ finitewave/cpuwave/model/courtemanche.py
@@ -1,4 +1,5 @@
-from numba import njit, prange
+import numpy as np
+from numba import njit, prange, typed
 
 from .cardiac_model import CardiacModel
 from ._registry import load_ops
@@ -55,7 +56,7 @@ class Courtemanche(CardiacModel):
 
     def __init__(self, memory_save=False):
         super().__init__(memory_save)
-        self.D_model = 0.154
+        self.D_model = 0.1544
         self.state_vars = ["u", "nai", "ki", "cai", "caup", "carel", "m", "h",
                            "j_", "d", "f", "oa", "oi", "ua", "ui", "xr", "xs",
                            "fca", "irel", "vrel", "urel", "wrel"]
@@ -144,6 +145,62 @@ class Courtemanche(CardiacModel):
                      self.kmcsqn, self.trpnmax, self.cmdnmax, self.csqnmax,
                      self.inacamax, self.inakmax, self.ipcamax, self.krel,
                      self.iupmax, self.kq10)
+        
+    def prepacing(self, stim_sequence):
+        stim_values = []
+        t_max = 0
+
+        for stim in stim_sequence:
+            n_beats = stim["n_beats"]
+            dt = stim["dt"]
+            bcl = stim["cycle_length"]
+            duration = stim["stim_duration"]
+            stim_amplitude = stim["stim_amplitude"]
+
+            stim_val = self._build_prepacing(dt, n_beats, bcl, duration, stim_amplitude)
+            stim_values.append(stim_val)
+            t_max += dt * len(stim_val)
+
+        stim_values = np.concatenate(stim_values)
+        self.u_pacing, state_vars = prepacing(
+            dt, t_max, stim_values, self.init_u,
+            self.init_nai, self.init_ki,
+            self.init_cai, self.init_caup, self.init_carel, self.init_m, self.init_h, self.init_j_,
+            self.init_d, self.init_f, self.init_oa, self.init_oi, self.init_ua, self.init_ui,
+            self.init_xr, self.init_xs, self.init_fca, self.init_irel, self.init_vrel,
+            self.init_urel, self.init_wrel,
+            self.gna, self.gnab, self.gk1, self.gkr, self.gks,
+            self.gto, self.gcal, self.gcab, self.gkur_coeff, self.F,
+            self.T, self.R, self.Vc, self.Vj, self.Vup, self.Vrel,
+            self.ibk, self.cao, self.nao, self.ko, self.caupmax,
+            self.kup, self.kmnai, self.kmko, self.kmnancx,
+            self.kmcancx, self.ksatncx, self.kmcmdn, self.kmtrpn,
+            self.kmcsqn, self.trpnmax, self.cmdnmax, self.csqnmax,
+            self.inacamax, self.inakmax, self.ipcamax, self.krel,
+            self.iupmax, self.kq10)
+        
+        # print(state_vars)
+        # initial conditions
+        for var, val in state_vars.items():
+            if var == "j":
+                var += "_"
+            setattr(self, "init_" + var, val)
+            
+
+    def _build_prepacing(self, dt, n_beats, bcl, stim_duration, stim_amplitude):
+        t_max = n_beats * bcl
+
+        stim_values = np.zeros(int(t_max / dt), dtype=np.float64)
+
+        for s in np.arange(n_beats):
+            stim_start = s * bcl
[…]

[…]
