### 2026-03-05 2aaad35
Update stimulations and solvers to make them consistent with ECG Tracker
