### 2026-03-07 #9310 @ashahide
Hi @piyueh - apologies for the slow turnaround time. I started a new role and have been settling in.

I completed my initial review. I think the paper and software are both very high quality, address an important gap in the literature, and overall will be a great addition.

I have identified three items that may potentially need to be updated before I can finish the review. However, I would appreciate if you could weigh in to ensure they are necessary to align with JOSS standards. While these are discussed, I will finish my review of the functional claims (preliminary results so far are in line with what the authors have shown). I expect to have the functional part of the review finished today.

Authorship
- Two of the authors do not seem to be listed as contributors in the git repo. Is this acceptable with JOSS standards?

Contributing
- There does not seem to be a documented statement of how users may contribute to the repository, report bugs, or seek support.

Documentation
- Pytest is not included in the list of dependencies, but is necessary to run the tests. Additionally, testing is not documented (e.g. how to run the tests), but I can attest that they all passed succesfully with by simply running pytest from the project root.

Edit:

I have finished reviewing the functional claims in the documentation and the paper. Two pipeline examples are given. Both ran successfully, with the paper example replicating the same figure.

Two UserWarnings were encounted that should be mitigated before I can check off the functionality part of the review. Neither of the examples produced a figure when plt.show() was executed, instead producing the following warning:

`UserWarning: FigureCanvasAgg is non-interactive, and thus cannot be shown`

Instead, the figures in my review were generated by removing the plt.show() and instead saving the figures with plt.savefig().

The authors should also verify that the documentation example (in the repo) produces the figure that I have included here.
