### 2026-03-09 #9310 @piyueh
@ashahide Thanks for helping with the JOSS review. If you were referring to the last two authors, the lead author provided an explanation here: https://github.com/openjournals/joss-reviews/issues/8800#issuecomment-3299628282

Per JOSS policy, non-coders may be listed as authors if they contributed directly to the project, including planning and decision-making.

Hi @TiNezlobinsky, could you address the reviewer’s comments regarding:

1. documentation for community contributions (reference: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines),
2. dependencies, testing, and documentation on how to run the tests, and
3. the plotting issues.

Thanks
