### 2026-03-10 d5d2839
Add LuoRudy91 model
