+++ finitewave/cpuwave/model/luo_rudy_91.py
@@ -1,0 +1,287 @@
+import numpy as np
+from numba import njit, prange, typed
+
+from .cardiac_model import CardiacModel
+from ._registry import load_ops
+from ._jitwrap import wrap_calc
+
+ops = load_ops("luo_rudy_91")
+jit_ops = wrap_calc(ops)
+
+calc_rhs = jit_ops["calc_rhs"]
+calc_dm = jit_ops["calc_dm"]
+calc_dh = jit_ops["calc_dh"]
+calc_dj = jit_ops["calc_dj"]
+calc_dd = jit_ops["calc_dd"]
+calc_df = jit_ops["calc_df"]
+calc_dx = jit_ops["calc_dx"]
+calc_dcai = jit_ops["calc_dcai"]
+calc_ina = jit_ops["calc_ina"]
+calc_isk = jit_ops["calc_isk"]
+calc_ik = jit_ops["calc_ik"]
+calc_ik1 = jit_ops["calc_ik1"]
+calc_ikp = jit_ops["calc_ikp"]
+calc_ib = jit_ops["calc_ib"]
+
+
+class LuoRudy91(CardiacModel):
+    """
+    Implements the Luo-Rudy 1991 ventricular action potential model.
+
+    This biophysically detailed model simulates the ionic currents and membrane potential 
+    of a ventricular cardiac cell based on Hodgkin-Huxley-type formalism. It was one of 
+    the first to incorporate realistic ionic channel kinetics, calcium dynamics, and 
+    multiple potassium currents to reproduce key phases of the action potential.
+
+    The model includes:
+    - Fast Na⁺ current (I_Na)
+    - Slow inward Ca²⁺ current (I_Si)
+    - Time-dependent K⁺ current (I_K)
+    - Time-independent K⁺ current (I_K1)
+    - Plateau K⁺ current (I_Kp)
+    - Background/leak current (I_b)
+
+    Attributes
+    ----------
+    state_vars : list of str
+        List of state variable names to save and restore (`u`, `m`, `h`, `j`, `d`, `f`, `x`, `cai`).
+    D_model : float
+        Diffusion coefficient representing electrical conductivity in the medium (typically set to 0.1).
+    gna, gsi, gk, gk1, gkp, gb : float
+        Maximum conductances for Na⁺, Ca²⁺, K⁺, and background channels [mS/μF].
+    ko, ki, nao, nai, cao : float
+        Ion concentrations in mM (extracellular and intracellular for Na⁺, K⁺, Ca²⁺).
+    R, T, F : float
+        Physical constants: gas constant, temperature in Kelvin, and Faraday constant.
+    PR_NaK : float
+        Sodium/potassium permeability ratio (used in reversal potential calculation for I_K).
+
+    Paper
+    -----
+    Luo CH, Rudy Y. 
+    A model of the ventricular cardiac action potential. Depolarization, repolarization, and their interaction. 
+    Circ Res. 1991 Jun;68(6):1501-26. 
+    doi: 10.1161/01.res.68.6.1501. 
+    PMID: 1709839.
+
+    """
+
+    def __init__(self, memory_save=False):
+        """
+        Initializes the LuoRudy91 instance, setting up the state variables and parameters.
+        """
+        super().__init__(memory_save)
+        self.D_model = 0.1
+
+        self.state_vars = ["u", "m", "h", "j", "d", "f", "x", "cai"]
+        self.npfloat = 'float64'
+
+        # initial conditions
+        params = ops.get_parameters()
+        for var, val in params.items():
+            setattr(self, var, val)
+
+        # initial conditions
+        variables = ops.get_variables()
+        for var, val in variables.items():
+            setattr(self, "init_" + var, val)
+
[…]
