+++ examples/models/luo_rudy_2d.py
@@ -31,46 +31,58 @@ import numpy as np
 import matplotlib.pyplot as plt
 import finitewave as fw
 
-n = 200
-m = 200
-# create mesh
-tissue = fw.CardiacTissue2D((n, m))
+n = 400
+m = 10
 
-# set up stimulation parameters
-stim_sequence = fw.StimSequence()
-stim_sequence.add_stim(fw.StimVoltageCoord2D(0, 1, 0, n, 0, 5))
-stim_sequence.add_stim(fw.StimVoltageCoord2D(180, 1, 0, n//2, 0, n))
+prepacing_protocol = [
+    {"n_beats": 30,
+     "cycle_length": 1000.,
+     "stim_duration": 1,
+     "stim_amplitude": 100.,
+     "dt": 0.01},
+    # {"n_beats": 100,
+    #  "cycle_length": 500.,
+    #  "stim_duration": 0.5,
+    #  "stim_amplitude": 100.,
+    #  "dt": 0.01}
+]
+# create mesh
+tissue = fw.CardiacTissueGrid((n, m), dr=0.1)
 
 # create model object and set up parameters
-luo_rudy = fw.LuoRudy912D()
-luo_rudy.dt = 0.01
-luo_rudy.dr = 0.25
-luo_rudy.t_max = 50
+luo_rudy = fw.LuoRudy91()
+luo_rudy.prepacing(prepacing_protocol)
+# set up stimulation parameters
+stim_sequence = fw.StimSequence()
+stim_sequence.add_stim(fw.StimCurrentCoord(0, 10., 1., 0, 1, 0, m))
 
-# add the tissue and the stim parameters to the model object
-luo_rudy.cardiac_tissue = tissue
-luo_rudy.stim_sequence = stim_sequence
-
-tracker_sequence = fw.TrackerSequence()
-action_pot_tracker = fw.ActionPotential2DTracker()
+action_pot_tracker = fw.ActionPotentialGridTracker()
 # to specify the mesh node under the measuring - use the cell_ind field:
 # eather list or list of lists can be used
-action_pot_tracker.cell_ind = [[50, 3]]
+action_pot_tracker.cell_ind = [[n//2, m//2]]
 action_pot_tracker.step = 1
-# tracker_sequence.add_tracker(action_pot_tracker)
-# luo_rudy.tracker_sequence = tracker_sequence
+
+tracker_sequence = fw.TrackerSequence()
+tracker_sequence.add_tracker(action_pot_tracker)
+
+simulation = fw.CardiacSimulation()
+simulation.dt = 0.01
+simulation.t_max = 5
+# add the tissue and the stim parameters to the model object:
+simulation.cardiac_tissue = tissue
+simulation.cardiac_model = luo_rudy
+simulation.stim_sequence = stim_sequence
+simulation.tracker_sequence = tracker_sequence
 
 # run the model:
-luo_rudy.run()
+simulation.run()
 
-plt.figure()
-plt.imshow(luo_rudy.u, cmap='RdBu_r')
-plt.colorbar()
+plt.plot(luo_rudy.u_pacing)
 plt.show()
 
 # # plot the action potential
 # plt.figure()
-# time = np.arange(len(action_pot_tracker.output)) * luo_rudy.dt
+# time = np.arange(len(action_pot_tracker.output)) * simulation.dt
 # plt.plot(time, action_pot_tracker.output, label="cell_50_3")
 # plt.legend(title='Luo-Rudy 1991')
 # plt.xlabel('Time (ms)')
