+++ finitewave/cpuwave/model/__init__.py
@@ -1,3 +1,4 @@
 from .aliev_panfilov import AlievPanfilov
 from .courtemanche import Courtemanche
 from .fenton_karma import FentonKarma
+from .luo_rudy_91 import LuoRudy91

[…]
