### 2026-03-10 def51aa
Fixing matplotlib agg backend issue
