### 2026-03-10 #9310 @TiNezlobinsky
Dear @ashahide,

Thank you for checking the functionality of the framework. I will try to address all the points you listed in your review.

1. **Matplotlib warning**

The warning is likely related to matplotlib backend used in headless environments, where plt.show() cannot display figures.

Our intention was to keep the examples simple and interactive, which is why we originally used plt.show(). However, your remark is indeed important.

As a compromise, we implemented the following check:
```python
if "agg" in plt.get_backend().lower():
    plt.savefig("figure.png", dpi=300)
else:
    plt.show()
```

This ensures that the figure is saved when running in a headless environment while still displaying it interactively when possible.

The script example and the README example have been updated accordingly.

2. **Contributing guidelines**

We added a CONTRIBUTING.md file describing how users can report issues, ask questions, and contribute to the project.

3. **Running tests**

Additional instructions on how to run the tests have been added to the README.

4. **Example verification**

We have also verified that the documentation example produces the expected figure.

Please let us know if anything else should be adjusted.
