+++ finitewave/cpuwave/model/luo_rudy_91.py
@@ -192,32 +192,29 @@ def ionic_kernel(u, rhs, indexes, dt,
     PR_NaK : float
         Sodium/potassium permeability ratio (used in reversal potential calculation for I_K).
     """
+    E_Na = (R * T / F) * np.log(nao / nai)
+    E_K1 = (R * T / F) * np.log(ko / ki)
+    
     for ind in prange(len(indexes)):
         ii = indexes[ind]
-
-        E_Na = (R * T / F) * np.log(nao / nai)
-        E_K1 = (R * T / F) * np.log(ko / ki)
-
-        m.flat[ii] += dt * calc_dm(u.flat[ii], m.flat[ii])
-        h.flat[ii] += dt * calc_dh(u.flat[ii], h.flat[ii])
-        j.flat[ii] += dt * calc_dj(u.flat[ii], j.flat[ii])
-
-        d.flat[ii] += dt * calc_dd(u.flat[ii], d.flat[ii])
-        f.flat[ii] += dt * calc_df(u.flat[ii], f.flat[ii])
-
-        x.flat[ii] += dt * calc_dx(u.flat[ii], x.flat[ii])
         
         ina = calc_ina(u.flat[ii], m.flat[ii], h.flat[ii], j.flat[ii], E_Na, gna)
         isi = calc_isk(u.flat[ii], d.flat[ii], f.flat[ii], cai.flat[ii], gsi)
-
-        cai.flat[ii] += dt * calc_dcai(cai.flat[ii], isi)
-
         ik = calc_ik(u.flat[ii], x.flat[ii], ko, ki, nao, nai, PR_NaK, R, T, F, gk)
         ik1 = calc_ik1(u.flat[ii], ko, E_K1, gk1)
         ikp = calc_ikp(u.flat[ii], E_K1, gkp)
         ib = calc_ib(u.flat[ii], gb)
 
         rhs.flat[ii] = calc_rhs(ina, isi, ik, ik1, ikp, ib)
+
+        m.flat[ii] += dt * calc_dm(u.flat[ii], m.flat[ii])
+        h.flat[ii] += dt * calc_dh(u.flat[ii], h.flat[ii])
+        j.flat[ii] += dt * calc_dj(u.flat[ii], j.flat[ii])
+        d.flat[ii] += dt * calc_dd(u.flat[ii], d.flat[ii])
+        f.flat[ii] += dt * calc_df(u.flat[ii], f.flat[ii])
+        x.flat[ii] += dt * calc_dx(u.flat[ii], x.flat[ii])
+        cai.flat[ii] += dt * calc_dcai(cai.flat[ii], isi)
+
 
 
 @njit
