### 2026-03-20 864b61a
Make IsotropicStencil second order accuracy
