+++ finitewave/cpuwave/numerics/fdm/isotropic_stencil.py
@@ -1,13 +1,85 @@
 import numpy as np
-from .stecil import Stencil
+from .stencil import Stencil
 
 
 class IsotropicStencil(Stencil):
     """
-    Assembles the isotropic stencil with first-order boundary conditions.
+    Isotropic stencil with second-order accuracy
     """
+
     def __init__(self):
         super().__init__()
+
+    def compute_diffusion_weights(self, mesh, diffusion, dr, indexes):
+        """
+        Computes the weights for the isotropic stencil with first-order
+        boundary conditions.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        diffusion : numpy.ndarray
+            The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        dr : float
+            The grid spacing.
+        indexes : numpy.ndarray
+            The indexes of the non-empty points in the mesh.
+
+        Returns
+        -------
+        tuple of np.ndarray
+            The rows, columns, and weights for the sparse matrix.
+        """
+        rows = []
+        cols = []
+        weights = []
+
+        ijk = np.array(np.unravel_index(indexes, mesh.shape))
+        for axis in range(mesh.ndim):
+            res = self.compute_flux_weights(mesh, diffusion, dr, ijk, axis)
+            r, c, w = self.compute_diffusion_component(mesh, dr, ijk, *res)
+            rows.append(r)
+            cols.append(c)
+            weights.append(w)
+
+        rows = np.concatenate(rows)
+        cols = np.concatenate(cols)
+        weights = np.concatenate(weights)
+        return rows, cols, weights
+    
+    def compute_diffusion_component(self, mesh, dr, ijk, ijk_list, w_list):
+        """
+        Computes the diffusion weights from the flux weights.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        dr : float
+            The grid spacing.
+        ijk : numpy.ndarray
+            The indexes of the non-empty points in the mesh.
+        ijk_list : list
+            The list of ijk coordinates of the involved points in the mesh.
+        w_list : list
+            The list of weights for the involved points.
+
+        Returns
+        -------
+        rows : np.ndarray
+            The row indexes for the sparse matrix.
+        cols : np.ndarray
+            The column indexes for the sparse matrix.
+        weights : np.ndarray
+            The weights for the sparse matrix.
+        """
+        rows, cols, weights = self.nonzero_weights(mesh, ijk, ijk_list, w_list)
+
+        rows = np.concatenate(rows)
+        cols = np.concatenate(cols)
+        weights = np.concatenate(weights) / dr
+        return rows, cols, weights
 
     def compute_flux_weights(self, mesh, diffusion, dr, ijk, axis):
         """
@@ -33,36 +105,30 @@ class IsotropicStencil(Stencil):
         tuple of np.ndarray
             The rows, columns, and weights for the flux computation.
         """
-        d_axis = diffusion[..., axis, axis]
-        ijk_neighbor = self.build_neighbor(ijk, shift=1, axis=axis)
-        valid_connection = (self.is_valid_index(ijk_neighbor, mesh) &
-                            self.is_valid_index(ijk, mesh))
-        weights = - self.average_diffusion(d_axis, ijk, ijk_neighbor, valid_connection) / dr
-        ijk_list = [ijk, ijk_neighbor]
-        w_list = [-weights, weights]
[…]
