### 2026-03-26 1c4b1e1
Update stencils and LAT Tracker
