+++ finitewave/cpuwave/numerics/fdm/asymmetric_stencil.py
@@ -1,5 +1,5 @@
 import numpy as np
-from .stecil import Stencil
+from .stencil import Stencil
 
 
 class AsymmetricStencil(Stencil):
@@ -33,7 +33,86 @@ class AsymmetricStencil(Stencil):
     """
     def __init__(self):
         super().__init__()
-   
+
+    def compute_diffusion_weights(self, mesh, diffusion, dr, indexes):
+        """
+        Computes the weights for the isotropic stencil with first-order
+        boundary conditions.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        diffusion : numpy.ndarray
+            The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        dr : float
+            The grid spacing.
+        indexes : numpy.ndarray
+            The indexes of the non-empty points in the mesh.
+
+        Returns
+        -------
+        rows : numpy.ndarray
+            The central node indexes.
+        cols : numpy.ndarray
+            The node indexes involved in diffusion calculation.
+        weights : numpy.ndarray
+            The weights for connections between central and neighboring nodes.
+        """
+        rows = []
+        cols = []
+        weights = []
+
+        ijk = np.array(np.unravel_index(indexes, mesh.shape))
+        for axis in range(mesh.ndim):
+            r, c, w = self.compute_diffusion_along_axis(mesh, diffusion, dr, ijk, axis)
+            rows.append(r)
+            cols.append(c)
+            weights.append(w)
+
+        rows = np.concatenate(rows)
+        cols = np.concatenate(cols)
+        weights = np.concatenate(weights)
+        return rows, cols, weights
+    
+    def compute_diffusion_along_axis(self, mesh, diffusion, dr, ijk, axis):
+        """
+        Computes the diffusion weights along a given axis.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        dr : float
+            The grid spacing.
+        ijk : numpy.ndarray
+            The indexes of the non-empty points in the mesh.
+        ijk_major : numpy.ndarray
+            The indexes of the major neighbor points.
+        ijk_list : list
+            A list of numpy arrays containing the indexes of the involved points.
+        w_list : list
+            A list of numpy arrays containing the flux weights for the involved points.
+
+        Returns
+        -------
+        rows : np.ndarray
+            The central node indexes.
+        cols : np.ndarray
+            The node indexes involved in diffusion calculation.
+        weights : np.ndarray
+            The weights for connections.
+        """
+        
+        ijk_major, ijk_list, w_list = self.compute_flux_weights(mesh, diffusion, dr, ijk, axis)
+        major_to_center = self.nonzero_weights(mesh, ijk, ijk_list, w_list)
+        center_to_major = self.nonzero_weights(mesh, ijk_major, ijk_list, w_list, direction=-1)
+
+        rows = np.concatenate(major_to_center[0] + center_to_major[0])
+        cols = np.concatenate(major_to_center[1] + center_to_major[1])
+        weights = np.concatenate(major_to_center[2] + center_to_major[2]) / dr
+        return rows, cols, weights
+    
     def compute_flux_weights(self, mesh, diffusion, dr, ijk, major_axis):
         """
         Computes the flux weights along major axis.
@@ -64,8 +143,12 @@ class AsymmetricStencil(Stencil):
 
         Returns 
         -------
-        tuple
[…]
