+++ finitewave/cpuwave/numerics/fdm/__init__.py
@@ -1,3 +1,4 @@
 from .asymmetric_stencil import AsymmetricStencil
 from .isotropic_stencil import IsotropicStencil
 from .cell_stencil import CellStencil
+from .symmetric_stencil import SymmetricStencil

[…]
