+++ examples/anatomical/atrial_elems.py
@@ -30,7 +30,7 @@ iir[iir > 1] = 1
 conductivity = 0.2 + 0.8 * iir
 
 coords = vtk_mesh.points / 1000
-coords /= 2
+coords /= 3
 elems = vtk_mesh.faces.reshape(-1, 4)[:, 1:4]
 fibers = vtk_mesh.cell_data['fiber_endo']
 conductivity = conductivity[elems].mean(axis=1)
@@ -53,62 +53,73 @@ tissue.fibers = fibers
 
 stim_indexes = np.random.choice(coords.shape[0], size=10, replace=False)
 stim_coords = coords[stim_indexes, :]
+
+stim_point_0 = coords[19600:19601]
+stim_point_1 = coords[72902:72903]
+
 # set up stimulation parameters:
 stim_sequence = fw.StimSequence()
 stim_sequence.add_stim(fw.StimCurrentElectrodes(0, 15, 0.1, geodesic.points, 1))
-# stim_sequence.add_stim(fw.StimCurrentElectrodes(0, 30, 0.1, coords[19600:19601], 1))
+stim_sequence.add_stim(fw.StimCurrentElectrodes(22, 15, 0.1, stim_point_1, 1))
 # stim_sequence.add_stim(fw.StimCurrentElectrodes(55, 15, 0.1, geodesic.points, 1))
-# stim_sequence.add_stim(fw.StimCurrentElectrodes(100, 15, 0.1, coords[72902:72903], 1))
-
+# stim_sequence.add_stim(fw.StimCurrentElectrodes(100, 15, 0.1, stim_point_1, 1))
 
 # for stim_time in [0, 26, 52, 78]:
 #     stim_sequence.add_stim(fw.StimCurrentElectrodes(stim_time, 30, 0.1, coords[33128:33129], 3))
 #     stim_sequence.add_stim(fw.StimCurrentElectrodes(stim_time, 30, 0.1, coords[30639:30640], 3))
 #     # stim_sequence.add_stim(fw.StimCurrentElectrodes(stim_time, 15, 0.1, coords[13372:13373], 1))
 
-
 state_point = coords[75398]
 dist = np.linalg.norm(coords - state_point, axis=1)
 state_indexes = np.where(dist < 8)[0]
 state_coords = coords[state_indexes]
 
+lat_tracker = fw.LocalActivationTimeTracker(step=10, start_time=70)
+tracker_sequence = fw.TrackerSequence()
+tracker_sequence.add_tracker(lat_tracker)
 
+
+prepacing_sequence = [
+    {"n_beats": 30,
+     "cycle_length": 30,
+     "stim_duration": 0.1,
+     "stim_amplitude": 20.,
+     "dt": 0.01},
+]
 cardiac_model = fw.AlievPanfilov()
-# cardiac_model.prepacing()
+cardiac_model.prepacing(prepacing_sequence)
 
 # create model object and set up parameters:
 simulation = fw.CardiacSimulation()
 simulation.dt = 0.01
-simulation.t_max = 1
+simulation.t_max = 100
 # simulation.state_loader = fw.StateLoader(path)
 # add the tissue and the stim parameters to the model object:
 simulation.cardiac_tissue = tissue
 simulation.cardiac_model = cardiac_model
-# simulation.stim_sequence = stim_sequence
+simulation.tracker_sequence = tracker_sequence
+simulation.stim_sequence = stim_sequence
 # simulation.stencil = stencil
 
-u = np.load(path / "u.npy")
-v = np.load(path / "v.npy")
+# u = np.load(path / "u.npy")
+# v = np.load(path / "v.npy")
 
 # u[state_indexes] = 0.
 # v[state_indexes] = 1.5
 
[…]
