+++ finitewave/cpuwave/__init__.py
@@ -2,6 +2,5 @@ from .simulation import *
 from .stimulation import *
 from .model import *
 from .numerics import *
-from .tools import *
 from .tissue import *
 from .tracker import *
