### 2026-03-30 fb3ab46
Remove Surface Assembler
