+++ finitewave/cpuwave/numerics/fem/surface_assembler.py
@@ -1,161 +1,0 @@
-import numpy as np
-
-from .element_assembler import VolumeAssembler
-
-
-class SurfaceAssembler(VolumeAssembler):
-    """
-    Class for assembling surface element diffusion models.
-
-    Attributes
-    ----------
-    reference_element : ReferenceElement
-        The reference element used for numerical integration.
-    """
-
-    def __init__(self):
-        super().__init__()
-
-    def compute_metrics(self, coords, elems):
-        """
-        Computes area and global gradients for surface elements in 3D
-        using the Jacobian.
-
-        Parameters:
-        ----------
-        coords: (N_nodes, 3)
-            Coordinates of the mesh nodes.
-        elems: (N_elems, N_points)
-            Element connectivity (node indices for each surface element).
-
-        Returns:
-        -------
-        areas: (N_elems,)
-            Area of each surface element.
-        grads: (N_elems, N_points, 3)
-            Gradient of shape functions in global coordinates for each element.
-        """
-        jacobian = self.build_jacobian(coords, elems)
-        areas = self._compute_areas(jacobian)
-        grads = self._compute_gradients(jacobian)
-
-        return areas, grads
-
-    def build_jacobian(self, coords, elems):
-        """
-        Build Jacobian matrices for surface elements.
-
-        Parameters:
-        ----------
-        coords: (N_nodes, 3)
-            Coordinates of the mesh nodes.
-        elems: (N_elems, N_points)
-            Element connectivity (node indices for each surface element).
-
-        Returns:
-        -------
-        jacobian: (N_elems, 2, 3)
-            Jacobian matrices for each surface element.
-        """
-        n_elems = elems.shape[0]
-        jacobian = np.zeros((n_elems, 2, 3))
-
-        for i in range(self.reference_element.n_points):
-            jacobian[:, 0, :] += (self.reference_element.dN_dxi[i] *
-                                  coords[elems[:, i]])
-            jacobian[:, 1, :] += (self.reference_element.dN_deta[i] *
-                                  coords[elems[:, i]])
-
-        return jacobian
-    
-    def compute_areas(self, coords, elems):
-        """
-        Computes areas of surface elements from their Jacobian matrices.
-
-        Parameters:
-        ----------
-        coords: (N_nodes, 3)
-            Coordinates of the mesh nodes.
-        elems: (N_elems, N_points)
-            Element connectivity (node indices for each surface element).
-
-        Returns:
-        -------
-            areas: (N_elems,)
-                Area of each surface element.
-        """
-        jacobian = self.build_jacobian(coords, elems)
-        areas = self._compute_areas(jacobian)
-        return areas
-
-    def _compute_gradients(self, jacobian):
-        """Compute global gradients for surface elements.
-
-        Parameters:
-        ----------
-        jacobian: (N, 2, 3)
-            Jacobian matrices for N surface elements.
-
-        Returns:
-        -------
-            grads: (N, n_points, 3)
-                Gradient of shape functions in global coordinates for each
-                element.
-
-        Note:
-        -----
-            The output shape is (N, n_points, 3) to match the expected format
-            for subsequent computations.
-        """
-        jacobian_inv = self.invert_jacobian(jacobian)
-        n_elems = jacobian_inv.shape[0]
-        n_points = self.reference_element.n_points
-        grads = np.zeros((n_elems, n_points, 3))
-
-        for i in range(n_points):
-            dN_ref = np.stack(
-                [np.full(n_elems, self.reference_element.dN_dxi[i]),
[…]
