### 2026-03-30 d4c2ab3
Add atrium animation
