### 2026-03-30 bc3c5c7
Merge kernel builder
