+++ finitewave/cpuwave/model/tp06_2d.py
@@ -1,1104 +1,0 @@
-import numpy as np
-from numba import njit, prange
-
-from finitewave.core.model.cardiac_model import CardiacModel
-from finitewave.cpuwave2D.stencil.asymmetric_stencil_2d import (
-    AsymmetricStencil2D
-)
-from finitewave.cpuwave2D.stencil.isotropic_stencil_2d import (
-    IsotropicStencil2D
-)
-
-
-class TP062D(CardiacModel):
-    """
-    Implements the ten Tusscher–Panfilov 2006 (TP06) human ventricular ionic model in 2D.
-
-    The TP06 model is a detailed biophysical model of the human ventricular 
-    action potential, designed to simulate realistic electrical behavior in 
-    tissue including alternans, reentrant waves, and spiral wave breakup.
-
-    This model includes:
-    - 18 dynamic state variables (voltage, ion concentrations, channel gates, buffers)
-    - Full calcium handling with subspace (cass) and sarcoplasmic reticulum (casr)
-    - Sodium, potassium, and calcium currents including background, exchanger, and pumps
-    - Buffering effects and intracellular transport
-
-    Finitewave provides this model in 2D form for efficient simulation and 
-    reproducible experimentation with custom spatial setups.
-
-    Attributes
-    ----------
-    D_model : float
-        Diffusion coefficient specific to this model (cm²/ms).
-    state_vars : list of str
-        List of all state variable names, used for checkpointing and logging.
-    ko : float
-        Extracellular potassium concentration (mM).
-    cao : float
-        Extracellular calcium concentration (mM).
-    nao : float
-        Extracellular sodium concentration (mM).
-    Vc : float
-        Cytoplasmic volume (μL).
-    Vsr : float
-        Sarcoplasmic reticulum volume (μL).
-    Vss : float
-        Subsarcolemmal space volume (μL).
-    R, T, F : float
-        Universal gas constant, absolute temperature, and Faraday constant.
-    RTONF : float
-        Precomputed RT/F value for Nernst equation.
-    CAPACITANCE : float
-        Membrane capacitance per unit area (μF/cm²).
-    gna, gcal, gkr, gks, gk1, gto : float
-        Conductances for major ionic channels.
-    gbna, gbca : float
-        Background sodium and calcium conductances.
-    gpca, gpk : float
-        Pump-related conductances.
-    knak, knaca : float
-        Maximal Na⁺/K⁺ pump and Na⁺/Ca²⁺ exchanger rates.
-    Km*, Kbuf*, Vmaxup, Vrel, etc.
-        Numerous kinetic constants for buffering, pump activity, and calcium handling.
-
-    Paper
-    -----
-    ten Tusscher KH, Panfilov AV. 
-    Alternans and spiral breakup in a human ventricular tissue model.
-    Am J Physiol Heart Circ Physiol. 2006 Sep;291(3):H1088–H1100.
-    https://doi.org/10.1152/ajpheart.00109.2006
-
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.D_model    = 0.154
-
-        self.state_vars = ["u", "cai", "casr", "cass", "nai", "Ki",
-                           "m", "h", "j", "xr1", "xr2", "xs", "r",
[…]
