### 2026-03-31 8d0c07d
Add text indentation for kernel generator
