### 2026-04-01 c477166
Add inified kernel generator
