### 2026-04-01 94682f2
Add Observer class
