### 2026-04-02 8bfb79d
Fix StepKernelGenerator
