+++ finitewave/cpuwave/model/_kernel_generators.py
@@ -1,7 +1,8 @@
 import textwrap
 import re
 import warnings
-import numpy as np
+
+from finitewave.cpuwave.model._kernel_builder import _build_cached
 
 from finitewave.core.model.kernel_generator import KernelGenerator
 
@@ -10,7 +11,7 @@ class StepKernelGenerator(KernelGenerator):
     def __init__(self):
         super().__init__()
         self.kernel_func_name = "ionic_kernel"
-        self.common_args = ["rhs", "indexes", "dt"]
+        self.common_args = ["rhs", "u", "indexes", "dt"]
     
     def _update_indexing(self, name, arrays):
         """
@@ -28,7 +29,7 @@ class StepKernelGenerator(KernelGenerator):
             return f"{name}.flat[idx]"
         return name
     
-    def generate_loop(self) -> str:
+    def generate_loop(self, indent) -> str:
         """
         Returns
         -------
@@ -36,12 +37,12 @@ class StepKernelGenerator(KernelGenerator):
             The header for the loop that iterates over the indexes.
         """
         loop = """\
-            for i in prange(len(indexes)):
+            for i in range(len(indexes)):
                 idx = indexes[i]
         """
-        return textwrap.dedent(loop).strip()
+        return textwrap.indent("\n" + textwrap.dedent(loop).strip(), " " * (indent))
     
-    def generate_assignments(self, arrays) -> str:
+    def generate_assignments(self, arrays, indent) -> str:
         """
         Returns
         -------
@@ -49,19 +50,22 @@ class StepKernelGenerator(KernelGenerator):
             Code for assigning arrays values to temporary (old) variables.
         """
         asign_vars = "\n".join(f"{var}_old = {self._assign_indexing(var, arrays)}" for var in arrays)
-        return textwrap.dedent(asign_vars).strip()
-    0
-    def generate_update_states(self, state_vars, arrays) -> str:
+        return textwrap.indent("\n" + textwrap.dedent(asign_vars).strip(), " " * (indent))
+    
+    def generate_update_states(self, state_vars, arrays, indent) -> str:
         """
         Returns
         -------
         str
             Code for updating state variables with new values after the step.
         """
-        update_vars = "\n".join(f"{self._update_indexing(var, arrays)} = {var}_new" for var in state_vars)
-        return textwrap.dedent(update_vars).strip()
+        update_vars = [var for var in state_vars if var != "u"]
+        update_vars += ["rhs"]
+
+        update_vars = "\n".join(f"{self._update_indexing(var, arrays)} = {var}_new" for var in update_vars)
+        return textwrap.indent("\n" + textwrap.dedent(update_vars).strip(), " " * (indent))
     
-    def generate_output(self, output_args) -> str:
+    def generate_output(self, output_args, indent) -> str:
         """
         Returns
         -------
@@ -69,183 +73,119 @@ class StepKernelGenerator(KernelGenerator):
             Code for writing the updated state variables back to the output arrays.
         """
         output = "return " + ", ".join(f"{var}" for var in output_args) if output_args else ""
-        return output
+        return textwrap.indent("\n" + textwrap.dedent(output).strip(), " " * (indent)) if output else ""
     
-    def generate_step(self, body, arrays, scalars, obs_args, state_vars) -> str:
-        input_args = ", ".join(f"{var}_old" for var in arrays)
-        input_args += (", " + ", ".join(f"{var}" for var in scalars))
-        output_args = "rhs_new, " + ", ".join(f"{var}_new" for var in state_vars)
+    def generate_step_func(self, step_func, arrays, scalars, state_vars, obs_args) -> str:
+
+        func_name, body = self.extract_func_body(step_func)
+
[…]
