+++ finitewave/core/model/__init__.py
@@ -1,1 +1,3 @@
 from .cardiac_model_base import CardiacModelBase
+from .observer import Observer
+from .kernel_generator import KernelGenerator

[…]
