### 2026-04-02 39948ce
Add prepacing kernel generator
