+++ finitewave/cpuwave/model/_cardiac_model.py
@@ -4,6 +4,7 @@ from warnings import warn
 from finitewave.core.model.cardiac_model_base import CardiacModelBase
 
 from ._kernel_generators import StepKernelGenerator
+from ._kernel_generators import SingleCellKernelGenerator
 
 
 class CardiacModel(CardiacModelBase):
@@ -48,7 +49,7 @@ class CardiacModel(CardiacModelBase):
         self.step = 1
         self.counter = 0
         self.ionic_kernel_generator = StepKernelGenerator()
-        self.single_cell_generator = None
+        self.prepacing_generator = SingleCellKernelGenerator("prepacing_kernel")
         self.observers = []
         self.model_func = {}
 
@@ -64,7 +65,60 @@ class CardiacModel(CardiacModelBase):
         self._allocate_arrays(simulation)
         self.rhs = np.zeros_like(self.u)
         self.compute_indexes(simulation.cardiac_tissue)
+
+    def run(self, dt):
+        """
+        Executes the ionic kernel for the Aliev-Panfilov model.
+        """
+        self.counter += 1
+        if (self.counter - 1) % self.step != 0:
+            return
+
+        self.ionic_kernel(
+            self.rhs,
+            self.u,
+            self.myo_indexes,
+            dt,
+            *self.ionic_kernel_args,
+        )
         
+    def _prepacing(self, stim_prepacing):
+        """Executes the prepacing sequence for the Aliev-Panfilov model."""
+
+        dt = stim_prepacing.dt
+        stim_values = stim_prepacing.stim_sequence
+        self.u_history = np.zeros(len(stim_values), dtype=self.npfloat)
+
+        prepacing_kernel_args = []
+        for name in self.prepacing_kernel_arg_names:
+            if name in self.state_vars:
+                param_val = getattr(self, f"init_{name}")
+
+                if not np.isscalar(param_val):
+                    raise ValueError(f"Prepacing kernel does not support array initial conditions. " +
+                                     f"Initial condition 'init_{name}' is an array.")
+                
+                prepacing_kernel_args.append(param_val)
+                continue
+            
+            if name in self.state_pars:
+                param_val = getattr(self, name)
+                if not np.isscalar(param_val):
+                    raise ValueError(f"Prepacing kernel does not support array parameters. " +
+                                     f"Parameter '{name}' is an array.")
+            
+                prepacing_kernel_args.append(getattr(self, name))
+                continue
+
+            raise ValueError(f"Prepacing kernel argument {name} not found in state variables or parameters.")
+
+        state_vals = self.prepacing_kernel(self.u_history, stim_values, dt,
+                                           self.init_u, *prepacing_kernel_args)
+
+        
+        # initial conditions
+        for var, value in zip(self.state_vars, state_vals):
+            setattr(self, f"init_{var}", value)
 
     def _initialize_variables_and_parameters(self, ops):
         self.default_parameters = ops.get_parameters()
@@ -112,7 +166,7 @@ class CardiacModel(CardiacModelBase):
                         f"param '{name}' shape {par.shape} != tissue shape {shape}"
                     ) 
     
-    def _initialize_ionic_kernel(self, ionic_step, model_func, exclude_params=[]):
+    def _initialize_ionic_kernel(self, ionic_step, exclude_params=[]):
 
         state_vars = self.state_vars
 
@@ -135,6 +189,40 @@ class CardiacModel(CardiacModelBase):
                                                    state_vars, self.model_func,
                                                    self.observers)
         self.ionic_kernel, self.ionic_kernel_arg_names = res
+
+    def _initialize_prepacing_kernel(self, ionic_step, exclude_params=[]):
+
+        state_vars = self.state_vars
+
+        arrays = ["u_history"]
+        scalars = []
+        
+        for param in self.state_pars:
+            if param in exclude_params:
[…]
