+++ finitewave/cpuwave/model/_kernel_generators.py
@@ -8,9 +8,9 @@ from finitewave.core.model.kernel_generator import KernelGenerator
 
 
 class StepKernelGenerator(KernelGenerator):
-    def __init__(self):
+    def __init__(self, kernel_func_name="ionic_kernel"):
         super().__init__()
-        self.kernel_func_name = "ionic_kernel"
+        self.kernel_func_name = kernel_func_name
         self.common_args = ["rhs", "u", "indexes", "dt"]
     
     def _update_indexing(self, name, arrays):
@@ -29,6 +29,20 @@ class StepKernelGenerator(KernelGenerator):
             return f"{name}.flat[idx]"
         return name
     
+    def _generate_args(self, vars, base="", suffix="", exclude=[]):
+
+        for var in vars:
+            if var in exclude:
+                continue
+
+            if base == "":
+                base += f"{var}{suffix}"
+                continue
+
+            base += f", {var}{suffix}"
+
+        return base
+
     def generate_loop(self, indent) -> str:
         """
         Returns
@@ -79,11 +93,17 @@ class StepKernelGenerator(KernelGenerator):
 
         func_name, body = self.extract_func_body(step_func)
 
-        input_args = 'dt, u, ' + ", ".join(f"{var}" for var in arrays if var not in self.common_args)
-        input_args += (", " + ", ".join(f"{var}" for var in scalars if var not in self.common_args))
+        input_args = "dt, u"
+        input_args = self._generate_args(arrays, input_args, exclude=self.common_args)
+        input_args = self._generate_args(scalars, input_args, exclude=self.common_args)
+
 
         output_args = "rhs, " + ", ".join(f"{var}_new" for var in state_vars if var not in self.common_args)
         output_args += (", " + ", ".join(obs_args) if obs_args else "")
+
+        output_args = "rhs"
+        output_args = self._generate_args(state_vars, output_args, suffix="_new", exclude=self.common_args)
+        output_args = self._generate_args(obs_args, output_args, exclude=self.common_args)
 
         body = textwrap.indent("\n" + textwrap.dedent(body).strip(), " " * (12 + 4))
 
@@ -96,11 +116,13 @@ class StepKernelGenerator(KernelGenerator):
 
         body_func = textwrap.dedent(body_func).strip()
 
-        signature_inputs = "dt, u_old, " + ", ".join(f"{var}_old" for var in arrays if var not in self.common_args)
-        signature_inputs += (", " + ", ".join(f"{var}" for var in scalars if var not in self.common_args))
+        signature_inputs = "dt, u_old"
+        signature_inputs = self._generate_args(arrays, signature_inputs, suffix="_old", exclude=self.common_args)
+        signature_inputs = self._generate_args(scalars, signature_inputs, exclude=self.common_args)
     
-        signature_returns = "rhs_new, " + ", ".join(f"{var}_new" for var in state_vars if var not in self.common_args)
-        signature_returns += (", " + ", ".join(obs_args) if obs_args else "")
+        signature_returns = "rhs_new"
+        signature_returns = self._generate_args(state_vars, signature_returns, suffix="_new", exclude=self.common_args)
+        signature_returns = self._generate_args(obs_args, signature_returns, exclude=self.common_args)
 
         func_signature = f"{signature_returns} = {func_name}({signature_inputs})"
         return func_name, func_signature, body_func
@@ -167,6 +189,11 @@ class StepKernelGenerator(KernelGenerator):
     def generate(self, step_func, arrays, scalars, state_vars, model_func, observers=[], output_args=[]):
[…]
