### 2026-04-08 90da159
Fix indexing problem with ElementAssembler
