+++ finitewave/cpuwave/numerics/fem/element_assembler.py
@@ -64,8 +64,17 @@ class ElementAssembler:
 
         for i in range(self.reference_element.n_points):
             for j in range(len(self.reference_element.dN)):
-                jacobian[:, j, :] += (self.reference_element.dN[j][i] *
+                jacobian[:, j, :] += (self.reference_element.dN[j, i] *
                                       coords[elems[:, i]])
+            
+        # for e in range(n_elems):
+        #     try:
+        #         np.linalg.inv(jacobian[e])
+        #     except np.linalg.LinAlgError:
+        #         print(f"Singular Jacobian for element {e}. Check the mesh quality.")
+        #         print(f"Element connectivity:\n{elems[e]}")
+        #         print(f"Element nodes:\n{coords[elems[e]]}")
+        #         raise ValueError()
         return jacobian
     
     def compute_gradients(self, coords, elems):
@@ -88,7 +97,7 @@ class ElementAssembler:
         jacobian = self.build_jacobian(coords, elems)
         grads = self._compute_gradients(jacobian)
         # Transpose to (N_elems, dim_phys, N_points)
-        grads = np.transpose(grads, (0, 2, 1)).copy() 
+        # grads = np.transpose(grads, (0, 2, 1)).copy() 
         return grads
     
     def compute_elements_size(self, coords, elems):
@@ -118,25 +127,25 @@ class ElementAssembler:
 
         Returns:
         -------
-            grads: (N_elems, N_points, dim_phys)
+            grads: (N_elems, dim_phys, N_points)
                 Gradient of shape functions in global coordinates for each
                 element.
 
         Note:
-            The output shape is (N_elems, N_points, dim_phys) to match the expected format
+            The output shape is (N_elems, dim_phys, N_points) to match the expected format
             for subsequent computations.
         """
         n_elems, dim_ref, dim_phys = jacobian.shape
         jacobian_inv = self.invert_jacobian(jacobian)
         n_points = self.reference_element.n_points
-        grads = np.zeros((n_elems, n_points, dim_phys))
+        grads = np.zeros((n_elems, dim_phys, n_points))
 
         for i in range(n_points):
             dN_ref = np.stack(
-                [np.full(n_elems, self.reference_element.dN[j][i]) for j in range(dim_ref)],
+                [np.full(n_elems, self.reference_element.dN[j, i]) for j in range(dim_ref)],
                 axis=1
             )
-            grads[:, i, :] = (jacobian_inv @ dN_ref[..., None])[..., 0]
+            grads[:, :, i] = (jacobian_inv @ dN_ref[..., None])[..., 0]
 
         return grads
 
@@ -193,6 +202,7 @@ class ElementAssembler:
     def compute_system_matrices(self, coords, elems, diffusion, indexes, reindex=False):
         """
         Computes the stiffness and mass matrices.
+
         Parameters
         ----------
         coords : np.ndarray
@@ -210,20 +220,67 @@ class ElementAssembler:
             The rows, columns, stiffness matrix, and mass matrix.
         """
 
-        print(diffusion.shape)
         shape = (coords.shape[0], coords.shape[0])
         elem_mass = self.reference_element.elem_mass
 
         if reindex:
             elems = self.reindex_elems(coords, elems, indexes)
+            coords = coords[indexes]
+            shape = (len(indexes), len(indexes))
 
         elements_size, grads = self.compute_metrics(coords, elems)
-        rows, cols, stiff, mass = stiffness_and_mass_matrix(elems, elements_size,
-                                                            grads, diffusion,
-                                                            elem_mass)
[…]
