+++ finitewave/cpuwave/model/__init__.py
@@ -3,6 +3,6 @@ from .aliev_panfilov import AlievPanfilov
 # from .mitchell_schaeffer import MitchellSchaeffer
 # from .fenton_karma import FentonKarma
 # from .bueno_orovio import BuenoOrovio
-# from .luo_rudy_91 import LuoRudy91
+from .luo_rudy_91 import LuoRudy91
 # from .ten_tusscher_panfilov_2006 import TenTusscherPanfilov2006
-# from .courtemanche import Courtemanche
+from .courtemanche import Courtemanche
