### 2026-04-09 0a3c711
Add node decoupling method for fibrosis modeling
