+++ finitewave/cpuwave/numerics/linalg/numba_linalg.py
@@ -2,6 +2,220 @@ import numpy as np
 from numba import njit, prange
 
 
+def forward_euler(A, u, rhs, u_new, indexes, dt):
+    """Performs the Forward Euler step for the diffusion equation.
+
+    For each time step:
+        1. Compute the diffusion term diff = A @ u.
+        2. Update the solution using u_new = u - diff + dt * rhs
+
+    Parameters
+    ----------
+    A : scipy.sparse.csr_matrix
+        The system matrix representing the diffusion operator.
+    u : np.ndarray
+        Current solution vector.
+    rhs : np.ndarray
+        Right-hand side vector (e.g., source term).
+    u_new : np.ndarray
+        Output solution vector to store the updated solution.
+    indexes : np.ndarray
+        Array of indexes where the solution is defined.
+    dt : float
+        Time step size.
+
+    Returns
+    -------
+    np.ndarray
+        Updated solution vector after the Forward Euler step.
+    """
+    indptr, indices, data = A.indptr, A.indices, A.data
+    return forward_euler_numba(indptr, indices, data, u, rhs, u_new, indexes, dt)
+
+
+@njit(parallel=True, fastmath=True, cache=True)
+def forward_euler_numba(indptr, indices, data, x, f, x_new, indexes, dt):
+    """
+    Performs the Forward Euler step:
+    x_new = x + A @ x + dt * f
+
+    Parameters
+    ----------
+    indptr : np.ndarray
+        CSR index pointer array.
+    indices : np.ndarray
+        CSR indices array.
+    data : np.ndarray
+        CSR data array.
+    x : np.ndarray
+        Current solution vector.
+    f : np.ndarray
+        Source term vector.
+    x_new : np.ndarray
+        Output solution vector.
+    indexes : np.ndarray
+        Array of indexes where the solution is defined.
+
+    Returns
+    -------
+    x_new : np.ndarray
+        Updated solution vector after the Forward Euler step.
+    """
+    n_rows = indptr.size - 1
+
+    for i in prange(n_rows):
+        start, end = indptr[i], indptr[i+1]
+        if start == end:
+            continue
+        a_x = 0.0
+        for j in range(start, end):
+            jj = indices[j]
+            jj = indexes[jj]
+            a_x += data[j] * x.flat[jj]
+
+        ii = indexes[i]
+        x_new.flat[ii] = x.flat[ii] - a_x + dt * f.flat[ii]
+
+    return x_new
+
+
+def cg_numba(A, b, x, indexes, rtol=None, atol=1e-6, maxiter=100):
+    """ Conjugate Gradient solver for Ax = b for x, where A is a sparse
+    matrix given in CSR format.
+
+    Parameters
+    ----------
+    A : scipy.sparse.csr_matrix
+        The matrix A in Ax = b.
+    b : 1D array of float
+        Right-hand side vector.
+    x : 1D array of float
+        Initial guess for the solution, will be modified in place.
+    indexes : 1D array of int
+        Array of indexes where the solution is defined.
+    atol : float
+        Absolute tolerance for the stopping criterion.
+    maxiter : int
+        Maximum number of iterations.
+
+    Returns
+    -------
+    x : 1D array of float
+        Approximate solution vector.
+    int
+        Number of iterations performed, or -1 if not converged within maxiter.
+    """
+    indptr, indices, data = A.indptr, A.indices, A.data
+    b_norm_2 = dot_numba(b, b, indexes)
+
+    if b_norm_2 == 0:
+        return x, 0
+    
+    if rtol is not None:
[…]
