+++ finitewave/cpuwave/simulation/diffusion/grid_diffusion_model.py
@@ -48,19 +48,18 @@ class GridDiffusionModel(DiffusionModelBase):
         model = self.simulation.cardiac_model
 
         mesh = tissue.mesh.copy()
-        mesh[mesh != 1] = 0
+        mesh[mesh != 1] = 0 
 
-        diffusion = self.compute_diffusion(mesh, tissue.conductivity,
-                                           tissue.fibers, tissue.D_al,
-                                           tissue.D_ac, model.D_model)
-        self.diffusion = diffusion.astype(self.simulation.npfloat)
-
-        self.weights = self.stencil.compute_system_matrices(mesh, self.diffusion,
+        self.diffusion = self.compute_diffusion_tensor(
+            mesh, tissue.conductivity, tissue.fibers, tissue.D_al, tissue.D_ac, model.D_model
+        )
+        self.connectivity_diffusion = self.convert_to_connectivity(self.diffusion, tissue.connectivity)
+        self.weights = self.stencil.compute_system_matrices(mesh, self.connectivity_diffusion,
                                                             tissue.dr, tissue.myo_indexes,
                                                             reindex=True)
         return self.weights
 
-    def compute_diffusion(self, mesh, conductivity, fibers, D_al, D_ac, D_model):
+    def compute_diffusion_tensor(self, mesh, conductivity, fibers, D_al, D_ac, D_model):
         """
         Computes the diffusion tensor based on fiber orientations.
 
@@ -90,6 +89,8 @@ class GridDiffusionModel(DiffusionModelBase):
             diffusion = np.zeros(mesh.shape + (ndim, ndim), dtype=np.float64)
             for i in range(ndim):
                 diffusion[..., i, i] = conductivity * D_model
+            
+            diffusion[mesh == 0, ...] = 0
             return diffusion
 
         ndim = fibers.shape[-1]
@@ -98,7 +99,56 @@ class GridDiffusionModel(DiffusionModelBase):
             for j in range(ndim):
                 d_ij = self.compute_diffusion_components(fibers, i, j, D_al, D_ac)
                 diffusion[..., i, j] = d_ij * conductivity * D_model
+
+        diffusion[mesh == 0, ...] = 0
         return diffusion
+    
+    def convert_to_connectivity(self, diffusion, connectivity):
+        """
+        Computes the diffusion tensor for junctions between myocytes (nodes).
+
+        Parameters
+        ----------
+        diffusion : np.ndarray (mesh.shape + (ndim, ndim))
+            The diffusion tensor for the myocytes.
+        connectivity : float or np.ndarray (mesh.shape + (ndim,))
+            The conductivity coefficient for the junctions between myocytes.
+
+        Returns
+        -------
+        np.ndarray (mesh.shape + (ndim, ndim))
+            The diffusion tensor for the junctions between myocytes.
+        """
+        connectivity_tensor = np.zeros_like(diffusion)
+
+        dim = diffusion.shape[-2]
+        
+        for d in range(dim):
+
+            if diffusion.shape[d] < 2:
+                continue
+            
+            index_left = [slice(None)] * (dim + 2)
+            index_left[d] = slice(None, -1)
+            index_left[dim] = d
+            index_left = tuple(index_left)
+            
+            index_right = [slice(None)] * (dim + 2)
+            index_right[d] = slice(1, None)
+            index_right[dim] = d
+            index_right = tuple(index_right)
+
+            d_i = (diffusion[index_left] + diffusion[index_right]) / 2.0
+            connectivity_tensor[index_left] = d_i
+        
+        connectivity_tensor[diffusion == 0] = 0
+
+        if np.isscalar(connectivity):
[…]
