### 2026-04-10 71b184d
Make variable trackers consistent with elements
