### 2026-04-12 a149307
Add phase map initialization
