+++ examples/grid/basic/3D/sphere_3d.py
@@ -104,9 +104,9 @@ stim_sequence.add_stim(stim1)
 stim_sequence.add_stim(stim2)
 
 stim_prepacing = fw.StimPrepacing(dt=0.01)
-stim_prepacing.add_stim(n_beats=5, basic_cycle_length=40, stim_duration=0.1, stim_amplitude=1)
-stim_prepacing.add_stim(n_beats=5, basic_cycle_length=30, stim_duration=0.1, stim_amplitude=1)
-stim_prepacing.add_stim(n_beats=5, basic_cycle_length=25, stim_duration=0.1, stim_amplitude=1)
+stim_prepacing.add_stim(n_beats=5, cycle_length40, stim_duration=0.1, stim_amplitude=1)
+stim_prepacing.add_stim(n_beats=5, cycle_length30, stim_duration=0.1, stim_amplitude=1)
+stim_prepacing.add_stim(n_beats=5, cycle_length25, stim_duration=0.1, stim_amplitude=1)
 
 model = fw.AlievPanfilov(memory_save=True)
 model.prepacing(stim_prepacing)
