+++ finitewave/core/model/cardiac_model_base.py
@@ -18,19 +18,32 @@ class CardiacModelBase(ABC):
         Array representing the sum of the ionic currents.
     D_model : float
         Model-specific diffusion coefficient.
+    step : int
+        Frequency of calling the ionic kernel (in terms of time steps).
+    iter_counter : int
+        Counter to track the number of simulation iterations.
     state_vars : list
-        List of state variables used during simulation.
+        List of model-specific state variables.
     state_pars : list
-        List of state parameters used during simulation.
+        List of model-specific parameters.
+    simulation : Simulation
+        Reference to the current simulation instance.
+    observers : list
+        List of observer functions to be called inside the `ionic_kernel`.
     """
     def __init__(self):
         self.u = None
         self.rhs = None
         self.D_model = None
 
+        self.step = 1
+        self.iter_counter = 0
+
         self.state_vars = []
         self.state_pars = []
         self.simulation = None
+
+        self.observers = []
 
     @abstractmethod
     def initialize(self, simulation):

[…]
