### 2026-04-13 e671384
Fix models. Add frame collecting option for FrameTracker
