+++ finitewave/cpuwave/model/__init__.py
@@ -1,8 +1,8 @@
 from .aliev_panfilov import AlievPanfilov
 # from .barkley import Barkley
 # from .mitchell_schaeffer import MitchellSchaeffer
-# from .fenton_karma import FentonKarma
-# from .bueno_orovio import BuenoOrovio
+from .fenton_karma import FentonKarma
+from .bueno_orovio import BuenoOrovio
 from .luo_rudy_91 import LuoRudy91
 # from .ten_tusscher_panfilov_2006 import TenTusscherPanfilov2006
 from .courtemanche import Courtemanche
