+++ finitewave/cpuwave/model/fenton_karma.py
@@ -1,24 +1,6 @@
-import math
-import numpy as np
-
-from finitewave.core.model.cardiac_model import CardiacModel
-from finitewave.core.model.ionic_kernel_generator import IonicKernelGenerator
-
-from finitewave.cpuwave.stencil.sten2D.asymmetric_stencil_2d import (
-    AsymmetricStencil2D
-)
-from finitewave.cpuwave.stencil.sten2D.isotropic_stencil_2d import (
-    IsotropicStencil2D
-)
-from finitewave.cpuwave.stencil.sten3D.asymmetric_stencil_3d import (
-    AsymmetricStencil3D
-)
-from finitewave.cpuwave.stencil.sten3D.isotropic_stencil_3d import (
-    IsotropicStencil3D
-)
+from ._cardiac_model import CardiacModel
 
 from finitewave.cpuwave.model._registry import load_ops, wrap_calc
-from finitewave.cpuwave.model._kernel_builder import build_kernel
 
 
 try:
@@ -29,36 +11,6 @@ except KeyError as e:
         "Fenton-Karma model ops not found. "
         # "Install model package: pip install finitewave-model-fenton-karma"
     ) from e
-
-
-class FentonKarmaKernel(IonicKernelGenerator):
-    def __init__(self):
-        super().__init__()
-        self.args_order = [
-            "u", "v", "w", "tau_d", "tau_o", "tau_r", "tau_si", "tau_v_m", "tau_v_p",
-            "tau_w_m", "tau_w_p", "k", "u_c", "uc_si"
-        ]
-
-    def generate_body(self) -> str:
-        model = {var: self._indexing(var) for var in (self.arrays + self.scalars)}
-        u_new = f"u_new{self._raw_indexing()}"
-        
-        return f"""\
-        J_fi = calc_Jfi({model['u']}, {model['v']}, 
-                            {model['u_c']}, {model['tau_d']})
-        J_so = calc_Jso({model['u']}, {model['u_c']},
-                            {model['tau_o']}, {model['tau_r']})
-        J_si = calc_Jsi({model['u']}, {model['w']},
-                            {model['k']}, {model['uc_si']}, {model['tau_si']})
-
-        {u_new} += dt * calc_rhs(J_fi, J_so, J_si)
-
-        {model['v']} += dt * calc_dv({model['v']} , {model['u']} , 
-                                          {model['u_c']} , {model['tau_v_m']} , {model['tau_v_p']} )
-        {model['w']}  += dt * calc_dw({model['w']} , {model['u']} , 
-                                          {model['u_c']} , {model['tau_w_m']} , {model['tau_w_p']})
-
-"""
 
 
 class FentonKarma(CardiacModel):
@@ -125,87 +77,11 @@ class FentonKarma(CardiacModel):
             
     """
 
-    def __init__(self):
+    def __init__(self, memory_save=False):
         """
         Initializes the Fenton-Karma instance with default parameters.
         """
-        super().__init__()
-        self.D_model = 1.
-        self.npfloat    = 'float64'
-
+        super().__init__(memory_save)
+        self.D_model = 0.1
         self._initialize_variables_and_parameters(ops)
-
-    def initialize(self):
-        """
-        Initializes the model for simulation.
-        """
-        super().initialize()
-
-        self._allocate_state_arrays()
-
-        gen = self._initialize_kernel(FentonKarmaKernel)
-        
-        glb = {
-            "calc_dv": jit_ops["calc_dv"], 
-            "calc_dw": jit_ops["calc_dw"],
-            "calc_Jfi": jit_ops["calc_Jfi"],
-            "calc_Jso": jit_ops["calc_Jso"],
[…]
