### 2026-04-14 22b2fe1
Optimize cg_numba
