### 2026-04-14 e687a50
Add option for FrameTracker
