### 2026-04-15 49a68ed
Intoduce mlx as alternative of numba for Apple M
