+++ finitewave/cpuwave/model/_kernel_generators.py
@@ -34,47 +34,7 @@ class IonicKernelGenerator(KernelGenerator):
         """
         super().__init__()
         self.kernel_func_name = kernel_func_name
-        self.common_args = ["rhs", "u", "indexes", "dt"]
-    
-    def _update_indexing(self, name, arrays):
-        """
-        Indexing that is used for updating state variables with new values.
-
-        Parameters
-        ----------
-        name : str
-            The name of the variable to update.
-        arrays : list of str
-            List of array variable names used in the step function.
-
-        Returns
-        -------
-        str
-            The indexing string to use for updating the variable.
-        """
-        if name in arrays:
-            return f"{name}.flat[idx]"
-        return name
-        
-    def _assign_indexing(self, name, arrays):
-        """
-        Indexing that is used to assign state variables to temporary (old) value.
-
-        Parameters
-        ----------
-        name : str
-            The name of the variable to assign.
-        arrays : list of str
-            List of array variable names used in the step function.
-
-        Returns
-        -------
-        str
-            The indexing string to use for assigning the variable.
-        """
-        if name in arrays:
-            return f"{name}.flat[idx]"
-        return name
+        self.common_args = ["dt", "u"]
     
     def _generate_args(self, vars, base="", suffix="", exclude=[]):
         """
@@ -151,7 +111,7 @@ class IonicKernelGenerator(KernelGenerator):
             The header for the loop that iterates over the indexes.
         """
         loop = """\
-            for i in range(len(indexes)):
+            for i in prange(len(indexes)):
                 idx = indexes[i]
         """
         return self._add_indent(loop, indent)
@@ -219,7 +179,7 @@ class IonicKernelGenerator(KernelGenerator):
         output = "return " + ", ".join(f"{var}" for var in output_args) if output_args else ""
         return self._add_indent(output, indent)
     
-    def generate_step_func(self, step_func, arrays, scalars, state_vars, obs_args):
+    def generate_step_func(self, step_func, state_vars, state_pars):
         """
         Generate the step function that computes the new state variables and rhs based on the current state and parameters.
         
@@ -227,15 +187,10 @@ class IonicKernelGenerator(KernelGenerator):
         ----------
         step_func : function
             The user-defined function that computes the new state variables and rhs.
-        arrays : list of str
-            List of array variable names used in the step function.
-        scalars : list of str
-            List of scalar variable names used in the step function.
         state_vars : list of str
             List of state variable names.
-        obs_args : list of str
-            List of observer extra arguments which should be returned from
-            the step function.
+        state_pars : list of str
+            List of state parameter names.
             
         Returns
         -------
@@ -250,12 +205,11 @@ class IonicKernelGenerator(KernelGenerator):
         func_name, body = self.extract_func_body(step_func)
 
         input_args = "dt"
-        input_args = self._generate_args(arrays, input_args, exclude=["rhs"])
-        input_args = self._generate_args(scalars, input_args)
+        input_args = self._generate_args(state_vars, input_args)
+        input_args = self._generate_args(state_pars, input_args)
                                                                                                                                                     
         output_args = "rhs"
         output_args = self._generate_args(state_vars, output_args, suffix="_new", exclude=["u"])
-        output_args = self._generate_args(obs_args, output_args, exclude=state_vars)
 
         body = self._add_indent(body, 16)
 
@@ -267,54 +221,17 @@ class IonicKernelGenerator(KernelGenerator):
         """
 
         func_body = textwrap.dedent(func_body).strip()
-
-        signature_inputs = "dt"
[…]
