### 2026-04-15 478a29b
Add mlxwave module
