+++ finitewave/mlxwave/numerics/fdm/stencil.py
@@ -1,0 +1,188 @@
+from abc import ABC, abstractmethod
+import numpy as np
+import scipy.sparse as sp
+
+
+class Stencil(ABC):
+    """
+    Base class for diffusion stencils.
+    """
+
+    def compute_system_matrices(self, mesh, diffusion, dr, indexes, reindex=False):
+        """
+        Computes the weights as a sparse matrix.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        diffusion : numpy.ndarray
+            The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        dr : float
+            The grid spacing.
+        indexes : numpy.ndarray
+            The indexes of the non-empty nodes in the mesh.
+
+        Returns
+        -------
+        scipy.sparse.csr_matrix
+            The stiffness matrix.
+        scipy.sparse.csr_matrix
+            The mass matrix. Diagonal matrix with ones on the diagonal.
+        """
+        rows, cols, weights = self.compute_diffusion_weights(mesh, diffusion, dr, indexes)
+        weights = weights.astype(diffusion.dtype)
+
+        size = mesh.size
+
+        if reindex:
+            rows, cols = self.reindex_matrix(mesh, rows, cols, indexes)
+            size = len(indexes)
+        
+        shape = (size, size)
+        # make stiffness matrix with positive diagonal
+        K_stiff = sp.csr_matrix((weights, (rows, cols)), shape=shape)
+        M_mass = sp.diags(np.ones_like(indexes, dtype=weights.dtype),
+                          offsets=0, format='csr')
+        return K_stiff.tocsr(), M_mass.tocsr()
+    
+    @abstractmethod
+    def compute_diffusion_weights(self, mesh, diffusion, dr, indexes):
+        """
+        Computes the weights for calculating the diffusion operator.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        diffusion : numpy.ndarray
+            The diffusion tensor as a (*mesh.shape, ndim, ndim).
+        dr : float
+            The grid spacing.
+        indexes : numpy.ndarray
+            The indexes of the non-empty nodes in the mesh.
+
+        Returns
+        -------
+        rows : numpy.ndarray
+            The central node indexes.
+        cols : numpy.ndarray
+            The node indexes involved in diffusion calculation.
+        weights : numpy.ndarray
+            The weights for connections between central and neighboring nodes.
+        """
+        raise NotImplementedError()
+
+    def nonzero_weights(self, mesh, ijk, ijk_list, w_list, direction=1):
+        """
+        Collects non-zero weights.
+
+        Parameters
+        ----------
+        mesh : numpy.ndarray
+            The mesh of the simulation.
+        ijk : numpy.ndarray
+            The indexes of the non-empty nodes in the mesh.
+        ijk_list : list
+            The list of ijk coordinates of the neighboring nodes in the mesh.
+        w_list : list
+            The list of weights for the neighboring nodes.
+        direction : int, optional
+            The direction of the flux (1 for major-to-center, -1 for center-to-major)
+
+        Returns
+        -------
+        rows : list
+            The list of row indexes for the sparse matrix.
+        cols : list
+            The list of column indexes for the sparse matrix.
+        weights : list
+            The list of weights for the sparse matrix.
+        """
+        rows, cols, weights = [], [], []
+
+        for i in range(len(w_list)):
+            w = w_list[i]
+            ijk_i = ijk_list[i]
+            rows += [np.ravel_multi_index(ijk[:, w != 0], mesh.shape)]
+            cols += [np.ravel_multi_index(ijk_i[:, w != 0], mesh.shape)]
[…]
