### 2026-04-15 df58024
Fix ForwardEulerSolverMlx
