### 2026-04-17 #9310 @cdaversin
I think the paper and software are suitable for publication. I especially emphasize the relevance of this software for teaching (intelligible and readable, with quite many examples and tutorials available). Here is my full review with comments and suggestions :

**Installation:**
I assume that  the branch `paper` should be used for this review, to match the implementation and usage example described in the paper. I have installed `FiniteWave` from source (branch `paper`) in a clean ubuntu:24.04 Docker container, running the commands provided in the README file. The installation worked as expected.
Suggested improvements for the installation section in the README:
- Indicate that `pip` is needed for the installation (`python3-pip` package)
- `build` is also needed (`pip install build`)

I notice that the current version of the software is available in PyPI, which is great. I also tried this way of installing, without problems.

The requirements (dependencies) are properly listed, with the required minimal versions. These dependencies are also included in the `pyproject.toml` with minimal versions, making the installation easy.
As already mentioned in the precedent review:
- `pytest` is missing in the list of dependencies, and in the .toml file.

**Documentation**:
	The README provides a fairly complete documentation describing the available features, illustrated by code samples. I have run the example provided in the README, it runs without problems and the returned figure is the same as the one shared by the first reviewer.
- It would be nice to add this figure to the README, for users running this example to check their installation.
- The README mentions the possibility to change the mesh to add fibrosis regions with dedicated marking of the nodes. It would be useful to add a code sample illustrating that, since there is no example for this, and the corresponding tests (`test_fibrosis_2d.py` and `test_fibrosis_2d.py`) or tutorials don’t seem to match the description from the README (no calls to `add_boundaries()` for example). Or update the README to be consistent with the examples.

The `paper` branch contains two PDF files as some sort of additional documentation. I don’t think these are necessary, and I see that these have been removed in the `main` branch.

**Paper**:
	The paper is well written and gives a nice overview of the software and features. The given example is reproducible with both the `paper` and `main` branches. And I was able to reproduce the figure shown in the paper with both versions.
Comments and suggestions :
- The reference for the figure (\autoref{fig:spiral_wave}) doesn’t show correctly on GitHub preview
- The authors cite a list of relevant software for cardiac electrophysiology, but these are described as “tightly coupled to HPC environments, and challenging to use and modify”. Such a statement is a bit strong in my opinion. Also, I would like to mention another relevant Python package performing cardiac electro-mechanics simulations (and pure electrophysiology simulations) in heart tissue: https://joss.theoj.org/papers/10.21105/joss.04753
- I am not sure what is meant by : “may lack multi-dimensional support”.
- The examples given in the README and the paper are based on a tissue slab, but is it possible to run the code on a bigger heart mesh like the one shown at the beginning of the README file ? (Is this GIF based on a FiniteWave simulation ?) And if so, is FiniteWave supporting parallel computing / can be run on clusters ?
