### 2026-04-17 3139a21
Apple Silicon GPU support
