+++ finitewave/cpuwave/tissue/__init__.py
@@ -1,7 +1,0 @@
-from .cardiac_tissue_grid import CardiacTissueGrid
-from .cardiac_tissue_elements import CardiacTissueElements
-
-__all__ = [
-    "CardiacTissueGrid",
-    "CardiacTissueElements",
-]
