+++ finitewave/core/stimul/__init__.py
@@ -1,0 +1,3 @@
+from finitewave.core.stimul.stim_sequence import StimSequence
+from finitewave.core.stimul.stim_prepacing import StimPrepacing
+from finitewave.core.stimul.stim_voltage_collection import StimVoltageCollection
