### 2026-04-19 b3b3498
fixes: readme, paper
