+++ finitewave/mlxwave/model/aliev_panfilov.py
@@ -1,15 +1,5 @@
 
 from finitewave.mlxwave.model._cardiac_model import CardiacModel
-from finitewave.mlxwave.model.kernel._load_ops import load_ops
-
-
-try:
-    ops = load_ops("aliev_panfilov")
-except KeyError as e:
-    raise ImportError(
-        "Aliev-Panfilov model ops not found. "
-        "Install model package: pip install finitewave-model-aliev-panfilov"
-    ) from e
 
 
 class AlievPanfilov(CardiacModel):
@@ -59,6 +49,8 @@ class AlievPanfilov(CardiacModel):
     https://doi.org/10.1016/0960-0779(95)00089-5.
     """
 
+    model_name = "aliev_panfilov"
+
     def __init__(self, memory_save=True):
         """
         Initializes the AlievPanfilov instance with default parameters.
@@ -71,5 +63,3 @@ class AlievPanfilov(CardiacModel):
         """
         super().__init__()
         self.D_model = 1.
-        self._initialize_variables_and_parameters(ops)
-        self.ops = ops
