### 2026-04-20 77a7c55
paper: references, readme: tests
