+++ paper/paper.bib
@@ -75,6 +75,34 @@
   doi = {10.1016/j.pbiomolbio.2015.12.008}
 }
 
+@article{Finsberg2023,
+  doi = {10.21105/joss.04753},
+  url = {https://doi.org/10.21105/joss.04753},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {81},
+  pages = {4753},
+  author = {Henrik Nicolay Topnes Finsberg and Ilsbeth Gerarda Maria van Herck and Cécile Daversin-Catty and Hermenegild Arevalo and Samuel Wall},
+  title = {simcardems: A FEniCS-based cardiac electro-mechanics solver},
+  journal = {Journal of Open Source Software}
+}
+
+@article{KABUS2026110088,
+  title = {Pigreads: The Python-integrated GPU-enabled reaction-diffusion solver using OpenCL for cardiac electrophysiology and other applications},
+  journal = {Computer Physics Communications},
+  volume = {323},
+  pages = {110088},
+  year = {2026},
+  issn = {0010-4655},
+  doi = {10.1016/j.cpc.2026.110088},
+  url = {https://www.sciencedirect.com/science/article/pii/S0010465526000706},
+  author = {Desmond Kabus and Hans Dierckx and Tim {De Coster}},
+  keywords = {Reaction-diffusion, Finite differences, Python, OpenCL, GPU, Electrophysiology},
+  abstract = {Pigreads is a streamlined Python module for efficient numerical solution of reaction-diffusion systems on graphics cards (GPU), with CPU fallback. It exposes a simple and straight-forward NumPy-compatible API. Users may employ built-in models – including electrophysiology examples – or supply custom reaction terms. Supported features include 0D-3D uniform Cartesian grids, no-flux and periodic boundary conditions, anisotropic diffusion, spatially varying diffusion and reaction, and localised source terms. The project is open-source, tested, documented, and distributed with examples and tutorials. PROGRAM SUMMARY Program Title: Pigreads CPC Library link to program files: https://doi.org/10.17632/5k8jjx74hc.1 Developer’s repository link: https://gitlab.com/pigreads/pigreads Licensing provisions: MIT Programming language: Python, OpenCL Nature of problem: Reaction-diffusion systems model phenomena across physics, chemistry and biology; their numerical solution can be computationally demanding and research codes are often ad-hoc and poorly documented. Solution method: The implementation uses finite-differences spatial discretisation and explicit forward-Euler time stepping. Performance-critical kernels run in OpenCL and are called from Python using PyOpenCL; data are represented with NumPy arrays. Additional comments including restrictions and unusual features: Pigreads targets reaction-diffusion problems with a focus on cardiac electrophysiology applications and emphasises simplicity and reproducibility over advanced numerical complexity. We adhere to best practices in scientific software development, including version control, testing, and continuous integration.}
+}
+
+
 @article{Nezlobinsky_multiparam,
   title = {Multiparametric analysis of geometric features of fibrotic textures leading to cardiac arrhythmias},
   author = {Nezlobinsky, T. and Okenov, A. and Panfilov, A. V.},
