+++ paper/paper.md
@@ -35,7 +35,7 @@ a strong technical background. Its modular structure, fast numerical core, and s
 
 
 # Statement of need
-Over the past decades, numerous tools have been developed to simulate normal and pathological cardiac activity, including electrical propagation, contraction, and anatomical features. As cardiac models have become more detailed and biophysically accurate, their computational cost and complexity have increased. Platforms like **OpenCARP**, **Chaste** or **Lifex-ep** [@Plank_openCARP; @Cooper_Chaste; @africa2022lifexcore] provide powerful and well-established frameworks for large-scale and physiologically detailed simulations. However, they are often designed with high-performance computing workflows in mind and may require substantial setup effort and domain-specific expertise, which can limit accessibility for users without a strong computational background. Other tools like **Myokit** [@Clerx2016Myokit] provide a lightweight and user-friendly environment, but mostly focus on single-cell simulations.
+Over the past decades, numerous tools have been developed to simulate normal and pathological cardiac activity, including electrical propagation, contraction, and anatomical features [@Plank_openCARP; @Cooper_Chaste; @africa2022lifexcore; @Clerx2016Myokit; @Finsberg2023; @KABUS2026110088]. As cardiac models have become more detailed and biophysically accurate, their computational cost and complexity have increased. Some existing frameworks provide powerful capabilities for large-scale and physiologically detailed simulations, but are often designed with high-performance computing workflows in mind and may require substantial setup effort and domain-specific expertise, which can limit accessibility for users without a strong computational background. Other tools provide a more lightweight and user-friendly environment, but tend to focus on reduced-complexity or single-cell workflows.
 
 **Finitewave** addresses this gap by offering a lightweight, transparent, and Python-native framework for cardiac modeling. It is designed to enable early user engagement in the modeling process, with a clear and modular structure that supports experimentation, learning, and customization.
 Its Python foundation allows smooth integration with other libraries (e.g., NumPy, Matplotlib, SciPy, Jupyter) and makes it ideal for use in educational and research settings.
