+++ pyproject.toml
@@ -25,13 +25,12 @@ dependencies = [
   "ffmpeg-python>=0.2.0",
   "pandas>=2.2.3",
   "scikit-image>=0.24.0",
+  "pytest",
+  "pytest-cov",
 ]
 
 [project.optional-dependencies]
-test = [
-  "pytest",
-  "pytest-cov"
-]
+
 docs = [
   "sphinx",
   "sphinx-rtd-theme",
@@ -40,14 +39,3 @@ docs = [
   "sphinx-copybutton",
   "numpydoc",
 ]
-
-# [tool.setuptools]
-# zip-safe = false
-# include-package-data = false
-# packages = [
-#    'finitewave',
-#    'finitewave.core',
-#    'finitewave.cpuwave2D',
-#    'finitewave.cpuwave3D',
-#    'finitewave.tools',
-# ]
