+++ finitewave_paper/pyvenv.cfg
@@ -1,0 +1,3 @@
+home = C:\Users\peppe\AppData\Local\Programs\Python\Python310
+include-system-site-packages = false
+version = 3.10.9
