### 2026-04-20 #9310 @TiNezlobinsky
Dear @cdaversin,

Thank you very much for your detailed review! We have addressed all comments below.

1. Installation

We added python `-m pip install --upgrade pip` at the beginning to install the latest pip package before installing the package.

`python -m build` was already in the Readme.

We always put `pytest` as an optional dependency in `pyproject.toml` - we have now moved it to a dependency section. This is fine for the paper branch.

2. Documentation

- Figure added to the Readme.
- We removed `add_boundaries()` from all the examples where it was, as calling it is no longer necessary.
- Fibrosis examples are in `examples/fibrosis`.
- We removed two PDF files from the paper branch.

3. Paper

- Fixed an issue with the image display in the preview. Should be OK now.
- Yes, you are right about that section. It can indeed be interpreted ambiguously. We softened the wording slightly and removed name references to specific packages. We also added new links to packages, including the one you suggested.

The new section now reads as follows:

> Over the past decades, numerous tools have been developed to simulate normal and pathological cardiac activity, including electrical propagation, contraction, and anatomical features [@Plank_openCARP; @Cooper_Chaste; @africa2022lifexcore; @Clerx2016Myokit; @Finsberg2023; @KABUS2026110088]. As cardiac models have become more detailed and biophysically accurate, their computational cost and complexity have increased. Some existing frameworks provide powerful capabilities for large-scale and physiologically detailed simulations, but are often designed with high-performance computing workflows in mind and may require substantial setup effort and domain-specific expertise, which can limit accessibility for users without a strong computational background. Other tools provide a more lightweight and user-friendly environment, but tend to focus on reduced-complexity or single-cell workflows.

- Sure. The GIF in the Readme was generated using Finitewave. The package supports 3D simulations on voxelized geometries, including anatomically shaped domains. The heart example is here: `Finitewave/examples/basics/3D/ventricle_geometry_3d.py`
- Parallel computing is available in the form of multithreading. Running on a cluster is possible, as long as it is a single node. We currently do not support the MPI protocol.

Thank you again for your time and review.
Timur
