### 2026-04-21 #9310 @cdaversin
Thanks @TiNezlobinsky  for the updates.

Regarding `pip`: what I meant was to add a sentence to the README indicating that one needs the `python3-pip` package to run the following `pip` commands. Especially since the software is meant to be accessible to students and users without a strong computational background.

Additional minor comment: In my opinion, the geometrical adaptability of the code (usable with anatomically shaped domains through voxelized geometries) deserves to be emphasized in the README / documentation, mentioning the 3D heart example.

Regardless, my comments and suggestions have all been taken into account, both in the README / documentation and in the paper.  @piyueh  My review is now complete, I think the software and paper are ready for publication.
Congratulations to the authors !
