### 2026-04-23 abdc967
Add jax support for CG solver
