+++ finitewave/cpuwave/solver/mlx_linalg/mlx_methods.py
@@ -50,16 +50,72 @@ class MlxCG:
     def __init__(self):
         pass
 
-    def solve(indices, data, b, x0=None, indexes=None, atol=1e-8, maxiter=100):
+    @staticmethod
+    @mx.compile
+    def axpy(a, x, y, indexes, out):
+        """
+        Performs the axpy operation:
+        out[indexes] = a * x[indexes] + y[indexes]
+
+        Parameters
+        ----------
+        a : float
+            Scalar multiplier.
+        x : 1D array of float
+            Vector to be scaled.
+        y : 1D array of float
+            Vector to be added.
+        indexes : 1D array of int
+            Array of indexes where the solution is defined.
+        out : 1D array of float
+            Output vector to store the result.
+
+        Returns
+        -------
+        np.ndarray
+            Updated solution vector after the axpy operation.
+        """
+        out[indexes] = a * x[indexes] + y[indexes]
+        return out
+    
+    @staticmethod
+    @mx.compile
+    def matvec(indices, data, x, indexes, out):
+        """
+        Performs A @ x where A is represented in ELLPACK format.
+
+        Parameters
+        ----------
+        indices : 2D array of int
+            Column indices corresponding to the non-zero values in data.
+            (global to the myocyte indexes, e.g. x[indices])
+        data : 2D array of float
+            Non-zero values of the matrix A in ELLPACK format.
+        x : 1D array of float
+            Vector to multiply with the matrix.
+        indexes : 1D array of int
+            Array of indexes where the solution is defined.
+
+        Returns
+        -------
+        np.ndarray (len(indexes),)
+            Result of the matrix-vector product.
+        """
+        out[indexes] = mx.sum(data * x[indexes][indices], axis=1)
+        return out
+    
+    @staticmethod
+    def solve(indices, data, b, x0, indexes, atol=1e-8, maxiter=100):
         """
         Conjugate Gradient solver for Ax = b where A is represented in ELLPACK format.
         
         Parameters
         ----------
-        data : 2D array of float
-            Non-zero values of the matrix A in ELLPACK format.
         indices : 2D array of int
             Column indices corresponding to the non-zero values in data.
+            (local to the myocyte indexes, e.g. x0[indexes][indices])
+        data : 2D array of float
+            Non-zero values of the matrix A in ELLPACK format.
         b : 1D array of float
             Right-hand side vector.
         x0 : 1D array of float, optional
@@ -79,19 +135,10 @@ class MlxCG:
         int
             Number of iterations performed, or -1 if not converged within maxiter.
         """
-
-        if x0 is not None:
-            x = x0
-        else:
-            x = mx.zeros_like(b)
-
-        if indexes is not None:
-            x_out = x
-            x = x[indexes]
-        
         # Initial residual and direction
         # r = b - A@x
-        r = b - mx.sum(data * x[indices], axis=1)
+        x = x0[indexes]
+        r = b[indexes] - mx.sum(data * x0[indexes][indices], axis=1)
         p = r
         r_norm = mx.sum(r * r)
         
@@ -105,15 +152,9 @@ class MlxCG:
                 if mx.sqrt(r_norm) < atol:
                     break
             
-        if indexes is not None:
-            x_out[indexes] = x
-        else:
-            x_out = x
-
-        if i == maxiter - 1:
-            return x_out, -1
+        x0[indexes] = x
             
-        return x_out, i
+        return x0, i
 
 
 @mx.compile
@@ -135,25 +176,4 @@ def cg_step(data, indices, x, r, p, r_norm):
     return x, r, p, r_norm_new
 
 
-@mx.compile
-def matvec_mlx(indices, data, x, indexes):
-    """
[…]
