### 2026-04-24 7c6aab6
Extract prepacing as a SingleCellModel
