+++ finitewave/cpuwave/solver/crank_nicolson_solver.py
@@ -36,7 +36,7 @@ class CrankNicolsonSolver(Solver):
         self.a_lhs_matrix = None
         self.a_rhs_matrix = None
         self.mass_matrix = None
-        self.method = None
+        self.linalg_method = None
 
     def initialize(self, simulation):
         """Initializes the Crank-Nicolson CG solver with the given simulation.
@@ -54,7 +54,8 @@ class CrankNicolsonSolver(Solver):
         self.rhs = simulation.cardiac_model.rhs
         self.myo_indexes = simulation.cardiac_model.myo_indexes
         self.assemble_system()
-        self.select_method(simulation.backend)
+        if self.linalg_method is None:
+            self.select_method(simulation.backend)
 
     def select_method(self, backend):
         if backend.name == "numba":
