### 2026-04-28 #9310 @piyueh
@TiNezlobinsky The next stage is my editorial review of the paper. Could you please complete the steps in "Additional Author Tasks After Review is Complete" [in the above comment](https://github.com/openjournals/joss-reviews/issues/9310#issuecomment-4336908770)? The version archived on Zenodo/figshare/etc. does not need to contain the paper itself (e.g., `paper.md`), but it must contain the source code being reviewed, including any changes made at the reviewers' request.
