### 2026-05-01 #9310 @TiNezlobinsky
@piyueh, done. Here is the DOI: https://doi.org/10.5281/zenodo.19946688.

Authors, orcids, license should be ok.
